==> analyses/06_tourist_accomodation_01.Rmd <==
---
title: "ABS tourist accomodation data 01"
author: "Radoslaw Panczak"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: 
  html_document: 
    highlight: pygments
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 4
    self_contained: true
editor_options: 
  chunk_output_type: console
knit: (function(inputFile, encoding) {
  rmarkdown::render(inputFile, encoding = encoding, output_dir = "../docs") })
---

<link rel="stylesheet" href="http://vis.supstat.com/assets/themes/dinky/css/scianimator.css">

<!-- ------------------------------------------------------------ --> 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE, fig.width = 8, fig.height = 6, dpi = 300, out.width = "800px", out.height = "600px")

knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

set.seed(12345)

library(tidyverse)
library(readxl)
library(lubridate)

library(rgdal)
library(rgeos)
library(tmap) # library(RColorBrewer)
tmap_mode("view") # makes map interactive
# library(sf)

library(statar)
library(animation)
```

```{r geodata_prep, eval=FALSE}
library(rmapshaper)

# SA2 outlines >> ACHTUNG >> 2011 !!!
SA2 <- readOGR("./Data/geo/ABS/raw/1270055001_sa2_2011_aust_shape", "SA2_2011_AUST", verbose = TRUE)
class(SA2)
nrow(SA2@data)
# View(SA2@data)
# qtm(SA2)
proj4string(SA2)
proj4string(SA2) <- NA_character_ # remove CRS information from lnd
proj4string(SA2) <- CRS("+init=epsg:4283")

# removing other teritories
SA2 <- SA2[SA2$SA4_CODE11 != 901, ]

SA2 <- ms_simplify(SA2, keep = 0.05, weighting = 0.7) # default settings

qtm(SA2)
class(SA2)
nrow(SA2@data)
proj4string(SA2)

# writeOGR(SA2, "./Data/geo/ABS/clean", "SA2_2011_AUST_clean", driver="ESRI Shapefile", overwrite_layer=TRUE)
saveRDS(SA2, file = "./Data/geo/ABS/clean/SA2_2011_AUST_clean.rds")
```


```{r geodata_load, include=FALSE}
# SA2 <- readOGR("./Data/geo/ABS/clean", "SA2_2011_AUST_clean", verbose = TRUE)
SA2 <- readRDS(file = "./Data/geo/ABS/clean/SA2_2011_AUST_clean.rds")
# SA2_f <- broom::tidy(SA2)

states <- readOGR("./Data/geo/ABS/clean", "STE_2016_AUST_clean", verbose = TRUE)
states_f <- broom::tidy(states)

# GCC <- readOGR("./Data/geo/ABS/clean", "GCCSA_2016_AUST_clean", verbose = TRUE)
# GCC_f <- broom::tidy(GCC)


# SA2_centr <- gCentroid(SA2, byid = TRUE)
#
# SA2_centr <- as.data.frame(gCentroid(SA2, byid = TRUE))
# colnames(SA2_centr) <- c("lon", "lat")
# SA2_centr <- data.frame("ID" = 1:nrow(SA2_centr), SA2_centr)
#
# # Create SpatialPointsDataFrame object
# coordinates(SA2_centr) <- c("lon", "lat")
# proj4string(SA2_centr) <- proj4string(SA2) # assign projection
# SA2_centr@data <- sp::over(x = SA2_centr, y = SA2, returnList = FALSE)
# SA2_centr@data$AREASQKM16 <- NULL

SA2_centr <- readOGR("./Data/geo/ABS/raw/1270055001_sa2_2011_aust_shape", "SA2_2011_AUST_centr_inside", verbose = TRUE)
class(SA2)
nrow(SA2@data)
# View(SA2_centr@data)
# qtm(SA2_centr)
proj4string(SA2_centr)
proj4string(SA2_centr) <- NA_character_ # remove CRS information from lnd
proj4string(SA2_centr) <- CRS("+init=epsg:4283")

# removing other teritories
SA2_centr <- SA2_centr[SA2_centr$SA4_CODE11 != 901, ]

qtm(SA2_centr)
# View(SA2_centr@data)
```

<!-- ------------------------------------------------------------ --> 

# Summary 

This is an overview of data preparation for '8635.0 - Tourist Accommodation, Australia, 2015-16' [dataset](http://www.abs.gov.au/AUSSTATS/abs@.nsf/Lookup/8635.0Main+Features12015-16?OpenDocument) from ABS.

# ToDo

- Torism Regions data?
- Link to Airbnb?
- Link to Census Night?
- Link to Visitor Surveys?
- Add denominator? ERP?

<!-- ------------------------------------------------------------ --> 


# Data preparation

## Raw data

Data are openly available. Unfortunately they can only be dowloaded in form of horribly designed spreadsheets. Note the empty cells! Comment for them says:

> *not available for publication but included in totals where applicable, unless otherwise indicated*

## Data processing SA2 level

- Monthly values of **Guest Nights Occupied** (GNO) were extracted
- **Empty** cells are kept as missing data
- **Date** is formated to 1st of the month
- **'Migratory - Offshore - Shipping'** SA2 is removed
- Data are linked to spatial boundaries - unfortunately **2011 SA2s** are used :/


```{r, eval=FALSE}
# ACT - slightly different structrure than others!
ACT_1 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/ACT.xls",
  sheet = "Table_6", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2015-07` = `no.__11`, `2015-08` = `no.__12`, `2015-09` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Australian Capital Territory") %>%
  filter(!grepl("Total", SA2)) %>%
  gather(`2015-07`, `2015-08`, `2015-09`, key = "date", value = "GNO")

ACT_2 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/ACT.xls",
  sheet = "Table_7", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2015-10` = `no.__11`, `2015-11` = `no.__12`, `2015-12` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Australian Capital Territory") %>%
  filter(!grepl("Total", SA2)) %>%
  gather(`2015-10`, `2015-11`, `2015-12`, key = "date", value = "GNO")

ACT_3 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/ACT.xls",
  sheet = "Table_8", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2016-01` = `no.__11`, `2016-02` = `no.__12`, `2016-03` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Australian Capital Territory") %>%
  filter(!grepl("Total", SA2)) %>%
  gather(`2016-01`, `2016-02`, `2016-03`, key = "date", value = "GNO")

ACT_4 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/ACT.xls",
  sheet = "Table_9", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2016-04` = `no.__11`, `2016-05` = `no.__12`, `2016-06` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Australian Capital Territory") %>%
  filter(!grepl("Total", SA2)) %>%
  gather(`2016-04`, `2016-05`, `2016-06`, key = "date", value = "GNO")

ACT <- rbind(ACT_1, ACT_2, ACT_3, ACT_4) %>%
  mutate(State_short = "ACT") %>%
  mutate(State_long = "Australian Capital Territory")

rm(ACT_1, ACT_2, ACT_3, ACT_4)

# NSW
NSW_1 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/NSW.xls",
  sheet = "Table_10", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2015-07` = `no.__11`, `2015-08` = `no.__12`, `2015-09` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "New South Wales") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2015-07`, `2015-08`, `2015-09`, key = "date", value = "GNO")

NSW_2 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/NSW.xls",
  sheet = "Table_11", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2015-10` = `no.__11`, `2015-11` = `no.__12`, `2015-12` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "New South Wales") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2015-10`, `2015-11`, `2015-12`, key = "date", value = "GNO")

NSW_3 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/NSW.xls",
  sheet = "Table_12", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2016-01` = `no.__11`, `2016-02` = `no.__12`, `2016-03` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "New South Wales") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2016-01`, `2016-02`, `2016-03`, key = "date", value = "GNO")

NSW_4 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/NSW.xls",
  sheet = "Table_13", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2016-04` = `no.__11`, `2016-05` = `no.__12`, `2016-06` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "New South Wales") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2016-04`, `2016-05`, `2016-06`, key = "date", value = "GNO")

NSW <- rbind(NSW_1, NSW_2, NSW_3, NSW_4) %>%
  mutate(State_short = "NSW") %>%
  mutate(State_long = "New South Wales")

rm(NSW_1, NSW_2, NSW_3, NSW_4)

# NT
NT_1 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/NT.xls",
  sheet = "Table_10", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2015-07` = `no.__11`, `2015-08` = `no.__12`, `2015-09` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Northern Territory") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2015-07`, `2015-08`, `2015-09`, key = "date", value = "GNO")

NT_2 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/NT.xls",
  sheet = "Table_11", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2015-10` = `no.__11`, `2015-11` = `no.__12`, `2015-12` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Northern Territory") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2015-10`, `2015-11`, `2015-12`, key = "date", value = "GNO")

NT_3 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/NT.xls",
  sheet = "Table_12", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2016-01` = `no.__11`, `2016-02` = `no.__12`, `2016-03` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Northern Territory") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2016-01`, `2016-02`, `2016-03`, key = "date", value = "GNO")

NT_4 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/NT.xls",
  sheet = "Table_13", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2016-04` = `no.__11`, `2016-05` = `no.__12`, `2016-06` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Northern Territory") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2016-04`, `2016-05`, `2016-06`, key = "date", value = "GNO")

NT <- rbind(NT_1, NT_2, NT_3, NT_4) %>%
  mutate(State_short = "NT") %>%
  mutate(State_long = "Northern Territory")

rm(NT_1, NT_2, NT_3, NT_4)

# Qld
Qld_1 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/Qld.xls",
  sheet = "Table_10", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2015-07` = `no.__11`, `2015-08` = `no.__12`, `2015-09` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Queensland") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2015-07`, `2015-08`, `2015-09`, key = "date", value = "GNO")

Qld_2 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/Qld.xls",
  sheet = "Table_11", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2015-10` = `no.__11`, `2015-11` = `no.__12`, `2015-12` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Queensland") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2015-10`, `2015-11`, `2015-12`, key = "date", value = "GNO")

Qld_3 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/Qld.xls",
  sheet = "Table_12", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2016-01` = `no.__11`, `2016-02` = `no.__12`, `2016-03` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Queensland") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2016-01`, `2016-02`, `2016-03`, key = "date", value = "GNO")

Qld_4 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/Qld.xls",
  sheet = "Table_13", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2016-04` = `no.__11`, `2016-05` = `no.__12`, `2016-06` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Queensland") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2016-04`, `2016-05`, `2016-06`, key = "date", value = "GNO")

Qld <- rbind(Qld_1, Qld_2, Qld_3, Qld_4) %>%
  mutate(State_short = "Qld") %>%
  mutate(State_long = "Queensland")

rm(Qld_1, Qld_2, Qld_3, Qld_4)


# SA
SA_1 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/SA.xls",
  sheet = "Table_10", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2015-07` = `no.__11`, `2015-08` = `no.__12`, `2015-09` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "South Australia") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2015-07`, `2015-08`, `2015-09`, key = "date", value = "GNO")

SA_2 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/SA.xls",
  sheet = "Table_11", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2015-10` = `no.__11`, `2015-11` = `no.__12`, `2015-12` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "South Australia") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2015-10`, `2015-11`, `2015-12`, key = "date", value = "GNO")

SA_3 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/SA.xls",
  sheet = "Table_12", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2016-01` = `no.__11`, `2016-02` = `no.__12`, `2016-03` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "South Australia") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2016-01`, `2016-02`, `2016-03`, key = "date", value = "GNO")

SA_4 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/SA.xls",
  sheet = "Table_13", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2016-04` = `no.__11`, `2016-05` = `no.__12`, `2016-06` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "South Australia") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2016-04`, `2016-05`, `2016-06`, key = "date", value = "GNO")

SA <- rbind(SA_1, SA_2, SA_3, SA_4) %>%
  mutate(State_short = "SA") %>%
  mutate(State_long = "South Australia")

rm(SA_1, SA_2, SA_3, SA_4)


# Tas
Tas_1 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/Tas.xls",
  sheet = "Table_10", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2015-07` = `no.__11`, `2015-08` = `no.__12`, `2015-09` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Tasmania") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2015-07`, `2015-08`, `2015-09`, key = "date", value = "GNO")

Tas_2 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/Tas.xls",
  sheet = "Table_11", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2015-10` = `no.__11`, `2015-11` = `no.__12`, `2015-12` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Tasmania") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2015-10`, `2015-11`, `2015-12`, key = "date", value = "GNO")

Tas_3 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/Tas.xls",
  sheet = "Table_12", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2016-01` = `no.__11`, `2016-02` = `no.__12`, `2016-03` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Tasmania") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2016-01`, `2016-02`, `2016-03`, key = "date", value = "GNO")

Tas_4 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/Tas.xls",
  sheet = "Table_13", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2016-04` = `no.__11`, `2016-05` = `no.__12`, `2016-06` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Tasmania") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2016-04`, `2016-05`, `2016-06`, key = "date", value = "GNO")

Tas <- rbind(Tas_1, Tas_2, Tas_3, Tas_4) %>%
  mutate(State_short = "Tas") %>%
  mutate(State_long = "Tasmania")

rm(Tas_1, Tas_2, Tas_3, Tas_4)


# Vic
Vic_1 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/Vic.xls",
  sheet = "Table_10", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2015-07` = `no.__11`, `2015-08` = `no.__12`, `2015-09` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Victoria") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2015-07`, `2015-08`, `2015-09`, key = "date", value = "GNO")

Vic_2 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/Vic.xls",
  sheet = "Table_11", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2015-10` = `no.__11`, `2015-11` = `no.__12`, `2015-12` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Victoria") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2015-10`, `2015-11`, `2015-12`, key = "date", value = "GNO")

Vic_3 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/Vic.xls",
  sheet = "Table_12", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2016-01` = `no.__11`, `2016-02` = `no.__12`, `2016-03` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Victoria") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2016-01`, `2016-02`, `2016-03`, key = "date", value = "GNO")

Vic_4 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/Vic.xls",
  sheet = "Table_13", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2016-04` = `no.__11`, `2016-05` = `no.__12`, `2016-06` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Victoria") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2016-04`, `2016-05`, `2016-06`, key = "date", value = "GNO")

Vic <- rbind(Vic_1, Vic_2, Vic_3, Vic_4) %>%
  mutate(State_short = "Vic") %>%
  mutate(State_long = "Victoria")

rm(Vic_1, Vic_2, Vic_3, Vic_4)


# WA
WA_1 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/WA.xls",
  sheet = "Table_10", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2015-07` = `no.__11`, `2015-08` = `no.__12`, `2015-09` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Western Australia") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2015-07`, `2015-08`, `2015-09`, key = "date", value = "GNO")

WA_2 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/WA.xls",
  sheet = "Table_11", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2015-10` = `no.__11`, `2015-11` = `no.__12`, `2015-12` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Western Australia") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2015-10`, `2015-11`, `2015-12`, key = "date", value = "GNO")

WA_3 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/WA.xls",
  sheet = "Table_12", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2016-01` = `no.__11`, `2016-02` = `no.__12`, `2016-03` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Western Australia") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2016-01`, `2016-02`, `2016-03`, key = "date", value = "GNO")

WA_4 <- read_excel("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/raw/WA.xls",
  sheet = "Table_13", skip = 6
) %>%
  dplyr::rename(SA2 = X__1, `2016-04` = `no.__11`, `2016-05` = `no.__12`, `2016-06` = `no.__13`) %>%
  filter(!is.na(`no.`)) %>%
  select(SA2, starts_with("20")) %>%
  filter(SA2 != "Western Australia") %>%
  filter(!grepl("Total", SA2)) %>%
  filter(!grepl("(TR)", SA2)) %>% # one TR name has no 'Total' keword :/
  gather(`2016-04`, `2016-05`, `2016-06`, key = "date", value = "GNO")

WA <- rbind(WA_1, WA_2, WA_3, WA_4) %>%
  mutate(State_short = "WA") %>%
  mutate(State_long = "Western Australia")

rm(WA_1, WA_2, WA_3, WA_4)


# All
TA_SA2_15_16 <- rbind(ACT, NSW, NT, Qld, SA, Tas, Vic, WA) %>%
  filter(!grepl("Migratory - Offshore - Shipping", SA2)) %>%
  mutate(date = parse_date_time(date, "ym")) %>%
  mutate(SA2 = ifelse(SA2 == "Central Highlands (TAS)", "Central Highlands", SA2))

rm(ACT, NSW, NT, Qld, SA, Tas, Vic, WA)

# table(TA_SA2_15_16$date)
# table(TA_SA2_15_16$State_short)

saveRDS(TA_SA2_15_16, file = "./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/clean/TA_SA2_15_16.rds")
write_csv(TA_SA2_15_16, "./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/clean/TA_SA2_15_16.csv")
```

```{r}
TA_SA2_15_16 <- readRDS("data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/clean/TA_SA2_15_16.rds")
```

# Results SA2 level

## Data structure

Final dataset consists of `r nrow(TA_SA2_15_16)` rows representing 12 observations (months) for `r n_distinct(TA_SA2_15_16$SA2)` SA2s.

There are `r sum(TA_SA2_15_16$GNO, na.rm = TRUE)` GNO in total.

## Missing data 

SA2s differ in how many months of data they have information on. 

```{r, include=FALSE}
missings <- TA_SA2_15_16 %>%
  group_by(SA2, State_short) %>%
  summarise(
    n = n(),
    no_value = sum(is.na(GNO)),
    missing = (no_value / n) * 100
  ) %>%
  mutate(missing = if_else(missing == Inf, 100, missing)) %>%
  ungroup()

missings %>%
  statar::sum_up(missing)

missings %>%
  filter(missing < 100) %>%
  statar::sum_up(missing)
```

There are `r nrow(filter(missings, missing < 100 ))` with any values, 
out of which `r nrow(filter(missings, missing == 0 ))` have complete year of monthly data and  
`r nrow(filter(missings, missing >= 100 ))` with no values. 
Additionally, there are other SA2s not present in the data which obviously will also have no values on accomodations. 

```{r, fig.width=4, fig.height=3, out.width="400px", out.height="300px"}
ggplot(missings, aes(no_value)) +
  geom_histogram(binwidth = 1) +
  theme_minimal(base_size = 8) +
  ggtitle("Missing data in Tourism Accomodation") +
  xlab("Number of months with missing data") +
  ylab("No. of SA2s")
```

Distribution of areas with and without data:

```{r, include=FALSE}
filled <- missings %>%
  filter(missing < 100) %>%
  select(SA2) %>%
  mutate(filled = 1)

SA2 <- left_join(SA2, filled, by = c("SA2_NAME16" = "SA2"))
SA2$filled[is.na(SA2$filled)] <- 0
table(SA2$filled)

zeroes <- missings %>%
  filter(missing >= 100) %>%
  select(SA2) %>%
  mutate(zeroes = 1)

SA2 <- left_join(SA2, zeroes, by = c("SA2_NAME16" = "SA2"))
SA2$zeroes[is.na(SA2$zeroes)] <- 0
table(SA2$zeroes)

table(SA2$filled, SA2$zeroes)

SA2$coverage <- NA
SA2$coverage <- ifelse(test = SA2$filled == 1, yes = "1 - data", no = SA2$coverage)
SA2$coverage <- ifelse(test = SA2$zeroes == 1, yes = "2 - no data", no = SA2$coverage)
SA2$coverage <- ifelse(test = SA2$zeroes == 0 & SA2$filled == 0, yes = "3 - excluded", no = SA2$coverage)

frq(SA2, coverage)
rm(missings)
# View(SA2@data)
```


```{r}
tm_shape(SA2) +
  tm_polygons("coverage", palette = "Pastel2", title = "TA data availability", border.col = "white", lwd = 0.5, id = "SA2_NAME11") +
  tm_shape(states) +
  tm_borders(lwd = 0.5, alpha = 0.5)
```

## Yearly totals

```{r, include=FALSE}
totals <- TA_SA2_15_16 %>%
  group_by(SA2) %>%
  summarise(
    n = n(),
    GNO = sum(GNO, na.rm = TRUE)
  ) %>%
  ungroup()

sum_up(totals, GNO)

SA2_centr <- sp::merge(SA2_centr, totals, by.x = "SA2_NAME11", by.y = "SA2")
rm(totals)
```

```{r}
tm_shape(states) +
  tm_borders(lwd = 0.5, alpha = 0.5) +
  tm_shape(SA2_centr) +
  tm_bubbles(
    size = "GNO", col = "darkgreen",
    border.col = "white", border.lwd = 1,
    alpha = 0.5,
    id = "SA2_NAME11"
  )
```

## Time trends

GNO per month by state:

```{r}
ggplot(data = TA_SA2_15_16, aes(x = date, y = GNO / 1000)) +
  geom_line(aes(group = SA2), alpha = 0.25, colour = "darkgreen") +
  facet_wrap(~State_short, ncol = 2) +
  theme_minimal(base_size = 10) +
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
  scale_colour_brewer(palette = "Set2") +
  labs(colour = "State") +
  xlab("") +
  scale_x_datetime(date_breaks = "3 months", date_minor_breaks = "1 month", date_labels = "%m-%y") +
  ggtitle("Guest nights occupied, SA2 level, 2015-2016")

# ggsave("\\\\gpem-atlas2/QCPR/Projects/ABS Linkage/Australian Population Studies/.pdf", width = 8, height = 4.5)
# ggsave("\\\\gpem-atlas2/QCPR/Projects/ABS Linkage/Australian Population Studies/.png", width = 8, height = 4.5)
```

Clear outliers of three biggest cities and Qld tourism regions dominate picture.


Aggregating to states: 

```{r}
group_by(TA_SA2_15_16, date, State_short) %>%
  summarise(GNO = sum(GNO, na.rm = TRUE)) %>%
  ungroup() %>%
  ggplot(aes(date, GNO / 1000, colour = State_short)) +
  geom_line(aes(group = State_short), size = 1) +
  theme_minimal(base_size = 10) +
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
  scale_colour_brewer(palette = "Set2") +
  labs(colour = "State") +
  xlab("") +
  scale_x_datetime(date_breaks = "3 months", date_minor_breaks = "1 month", date_labels = "%m-%y") +
  ggtitle("Guest nights occupied, State aggregates, 2015-2016")

# ggsave("\\\\gpem-atlas2/QCPR/Projects/ABS Linkage/Australian Population Studies/.pdf", width = 8, height = 4.5)
# ggsave("\\\\gpem-atlas2/QCPR/Projects/ABS Linkage/Australian Population Studies/.png", width = 8, height = 4.5)
```

## Space-Time trends 1 

```{r, fig.width=6, fig.height=8, out.width="600px", out.height="800px"}
left_join(TA_SA2_15_16, select(SA2_centr@data, SA2_NAME11, POINT_X, POINT_Y), by = c("SA2" = "SA2_NAME11")) %>%
  mutate(DM = format(date, "%Y-%m")) %>%
  ggplot() +
  geom_point(aes(POINT_X, POINT_Y, size = GNO / 1000), fill = "darkgreen", alpha = 0.66, colour = "white", pch = 21) +
  geom_polygon(data = states_f, aes(long, lat, group = group), fill = NA, colour = "grey40") +
  theme_void(base_size = 10) +
  ylab("") +
  xlab("") +
  coord_equal() +
  facet_wrap(~DM, ncol = 3) +
  ggtitle("Guest nights occupied, SA2 level, 2015-2016")

# ggsave("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/clean/GNO_facets.pdf", width = 6, height = 8)
# ggsave("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/clean/GNO_facets.png", width = 6, height = 8)
```

## Space-Time trends 2

```{r, include=FALSE}
temp <- left_join(TA_SA2_15_16, select(SA2_centr@data, SA2_NAME11, POINT_X, POINT_Y), by = c("SA2" = "SA2_NAME11")) %>%
  mutate(DM = format(date, "%Y-%m"))

sum_up(temp, GNO)

s <- seq(
  from = as.Date("2015-07-01", format = "%Y-%m-%d"),
  to = as.Date("2016-06-01", format = "%Y-%m-%d"),
  by = "months"
)
```

```{r,fig.width=6, fig.height=8, dpi=300, fig.show='animate', fig.path="img//", message=FALSE, warning=FALSE}
for (i in seq_along(s)) {
  myMap <- ggplot() +
    geom_point(
      data = filter(temp, as.Date(date, format = "%Y-%m-%d") == s[[i]]),
      aes(POINT_X, POINT_Y, size = GNO / 1000), fill = "darkgreen", alpha = 0.66, colour = "white", pch = 21
    ) +
    geom_polygon(data = states_f, aes(long, lat, group = group), fill = NA, colour = "grey40") +
    theme_void(base_size = 10) +
    ylab("") +
    xlab("") +
    coord_equal() +
    scale_size(name = "GNO", limits = c(1, 720), range = c(1, 10)) +
    ggtitle(paste0("Guest nights occupied in SA2 regions, ", format(s[[i]], "%Y-%m")))
  print(myMap)
}
```

```{r, include=FALSE}
rm(s, myMap, temp)
```
