==> analyses/06_tourist_accomodation_03.R <==
TA_SA2_15_16 <- readRDS("./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/clean/TA_SA2_15_16.rds")

TA_TR_GNO_15_16 <- readRDS(file = "./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/clean/TA_TR_GNO_15_16.rds")

TA_TR_GNO_15_16_all <- readRDS(file = "./Data/ABS/8635.0 Tourist Accommodation, Australia, 2015-16/clean/TA_TR_GNO_15_16_all.rds")
