==> analyses/08_Airbnb_03.Rmd <==
---
title: "Airbnb data 03"
author: "Radoslaw Panczak"
date: "`r format(Sys.time(), '%d %B, %Y')`"
mainfont: DejaVu Sans
output: 
  html_document: 
    # css: custom.css
    toc: true
    toc_depth: 4
    toc_float: true
    number_sections: true
    theme: united 
    highlight: pygments 
---

<!-- ------------------------------------------------------------ --> 
<!-- ------------------------------------------------------------ --> 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE, fig.width=8, fig.height=6, dpi=300, out.width="800px", out.height="600px")

# knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())

set.seed(12345)
options(scipen = 999)

library(pacman) 
p_load(tidyverse, magrittr, sf, tmap, skimr, summarytools, kableExtra, naniar, janitor)

tmap_mode("view") # makes map interactive

```

```{r geodata_prep, eval=FALSE}
p_load(rmapshaper)

# TR 2015 outlines 
TR15 <- st_read("data/geo/1270055003_tr_2015_aust_shape/TR_2015_AUST.shp", stringsAsFactors = FALSE)
class(TR15); nrow(TR15); names(TR15)
# View(TR15)
st_crs(TR15)
TR15$AREA_SQKM <- NULL

TR15 <- ms_simplify(TR15, keep = 0.05, weighting = 0.7) # default settings
# qtm(TR15)

# table(duplicated(TR15$TR_NAME15))

saveRDS(TR15, file = "data/geo/TR_2015_AUST.rds")

# TR 2016 outlines 
TR16 <- st_read("data/geo/1270055003_tr_2016_aust_shape/TR_2016_AUST.shp", stringsAsFactors = FALSE)
class(TR16); nrow(TR16); names(TR16)
# View(TR16)
st_crs(TR16)
TR16$AREASQKM16 <- NULL

TR16 <- ms_simplify(TR16, keep = 0.05, weighting = 0.7) # default settings
# qtm(TR16)

# table(duplicated(TR16$TR_NAME16))

saveRDS(TR16, file = "data/geo/TR_2016_AUST.rds")

# SA3 outlines
SA3 <- st_read("./data/geo/1270055001_sa3_2016_aust_shape/SA3_2016_AUST.shp", stringsAsFactors = FALSE)
# View(SA3)
# st_crs(SA3)
# qtm(SA3)
saveRDS(SA3, file = "./data/geo/SA3_2016_AUST.rds")

# removing other teritories and zeroes
SA3 <- 
  SA3 %>% 
  filter(SA4_CODE16 != 901) %>% 
  filter(AREASQKM16 != 0) %>% 
  st_transform(3112) 

nrow(SA3)

SA3 <- ms_simplify(SA3, keep = 0.05, weighting = 0.7) # default settings
# SA3 <- ms_simplify(SA3, keep = 0.04, weighting = 0.8) # bit more aggressive

nrow(SA3)
# qtm(SA3)

saveRDS(SA3, file = "./data/geo/SA3_2016_AUST_clean.rds")

# Mesh blocks of 'Blue Mountains - South' SA3 have 0 pops
# replaced with 'normal' centroid

SA3_centr <- read_csv("data/geo/SA3_2016_pop_center_median.csv") %>% 
  select(-OBJECTID) %>% 
  rename(lon = XCoord, 
         lat = YCoord) %>% 
  mutate(SA3_CODE16 = as.character(SA3_CODE16)) %>% 
  right_join(select(st_drop_geometry(SA3), SA3_CODE16)) %>% 
  mutate(lon = ifelse(SA3_CODE16 == "12402", 150.287952, lon)) %>% 
  mutate(lat = ifelse(SA3_CODE16 == "12402", -33.993425, lat)) %>% 
  st_as_sf(coords = c("lon", "lat"), crs = 4326, remove = FALSE) %>% 
  st_transform(3112) 

# qtm(SA3_centr)
 
saveRDS(SA3_centr, file = "./data/geo/SA3_centr.rds")

p_unload(rmapshaper)
```

```{r joins_prep, eval=FALSE, include=FALSE}
SA1_2016_join_new <- read_csv("./data/airdna/raw/Spatial_join/SA1_2016_join_new.zip") %>% 
  dplyr::rename(property_id = PROPERTY_ID) %>% 
  select(-XCoord, -YCoord, -JOIN_COUNT)

saveRDS(SA1_2016_join_new, file = "./data/airdna/clean/SA1_2016_join_new.rds")

POA_2016_join_new <- read_csv("./data/airdna/raw/Spatial_join/POA_2016_join_new.zip") %>% 
  select(PROPERTY_ID, POA_CODE16) %>% 
  dplyr::rename(property_id = PROPERTY_ID) # %>% mutate(POA_CODE16 = as.numeric(POA_CODE16))

saveRDS(POA_2016_join_new, file = "./data/airdna/clean/POA_2016_join_new.rds")
```

```{r, include=FALSE}
SA1_2016_join_new <- readRDS("./data/airdna/clean/SA1_2016_join_new.rds") %>% 
  dplyr::select(property_id, starts_with("SA1"), starts_with("SA2"), 
                starts_with("SA3"), starts_with("SA4"), 
                starts_with("STE"))
```

<!-- ------------------------------------------------------------ --> 
<!-- ------------------------------------------------------------ --> 

# Summary 

This is an overview of second **raw** `Airdna` dataset. 

Separate preparation is done for 'properties' and 'monthly_new' datasets. 

<!-- ------------------------------------------------------------ --> 
<!-- ------------------------------------------------------------ --> 

# Properties

```{r, include=FALSE}
property_new_raw <- read_csv("./data/airdna/raw/Australia_Airbnb_property_2019-02-01.zip")

# any(duplicated(property_new_raw$`property_new ID`))
# length(unique(property_new_raw$`property_new ID`))
# any(duplicated(property_new_raw$`Host ID`))
# length(unique(property_new_raw$`Host ID`))
```


## General

```{r, include=FALSE}
property_new <- property_new_raw %>% 
  clean_names() %>%
  remove_empty(c("rows", "cols")) %>% 
  select(-country) %>% 
  mutate(latitude = ifelse(property_id == 1476936, latitude*-1, latitude))

# property_new %<>% left_join(SA1_2016_join_new)

# table(is.na(SA1_2016_join_new$SA1_MAIN16))
# table(is.na(property_new$SA1_MAIN16))

# temp <- property_new %>% 
#   filter(is.na(SA1_MAIN16))

rm(property_new_raw) #, temp)
gc()

saveRDS(property_new, "./data/airdna/clean/property_new.Rds")
```

* All variable names were converted to lower caps
* Dots/gaps in the names were replaced by single underscores.
* One location with wrong coordinates (`property_id` == 1476936) was fixed (wrong hemmisphere!).
* Completely empty variables (`Zipcode` and `Metropolitan_Statistical_Area`) were dropped.
* Variable `Country` was droped. 
* Data was spatially linked to ABS geography with SA1 and above level variables added.

Raw dataset consists of `r nrow(property_new)` records of Austrlian Airbnb properties as scraped by Airdna. 

Dataset consists of `r ncol(property_new)` variables:

```{r}
names(property_new)
```

## Data structure

```{r, results='asis'}
# skim(property_new)
skim(property_new) %>% skimr::kable() 
```

## Missing data

```{r}
gg_miss_var(property_new, show_pct = TRUE)
```

## Listing Type

```{r, fig.width=4, fig.height=3, out.width="400px", out.height="300px"}
# tab(property_new, Listing_Type)
# tab(property_new, property_type, Listing_Type)
# filter(property_new, !is.na(Listing_Type)) %>% 
# ggplot(aes(property_type))+ 
#   geom_bar() +
#   coord_flip() + 
#   facet_wrap(~Listing_Type, ncol = 3)
ggplot(property_new, aes(listing_type))+ 
  geom_bar() + coord_flip() + theme_minimal()
```

## Created Date

```{r}
ggplot(property_new, aes(created_date))+ 
  geom_bar() + theme_minimal()
```

## Occupancy Rate

```{r, fig.width=4, fig.height=3, out.width="400px", out.height="300px"}
property_new %>% 
  select(occupancy_rate_ltm) %>% 
  summarytools::descr(stats = c("mean", "med", "min", "max")) 

ggplot(property_new, aes(occupancy_rate_ltm)) +
  geom_histogram() + theme_minimal()
```

## Bedrooms

```{r, fig.width=4, fig.height=3, out.width="400px", out.height="300px"}
property_new %>% 
  select(bedrooms) %>% 
  summarytools::descr(stats = c("mean", "med", "min", "max")) 

ggplot(property_new, aes(bedrooms)) +
  geom_histogram(binwidth = 1) + theme_minimal()
```

## Max Guests

```{r, fig.width=4, fig.height=3, out.width="400px", out.height="300px"}
property_new %>% 
  select(max_guests) %>% 
  summarytools::descr(stats = c("mean", "med", "min", "max")) 

ggplot(property_new, aes(max_guests)) +
  geom_histogram(binwidth = 1) + theme_minimal()
```


```{r}
property_new_select <- property_new %>% 
  select(property_id, host_id, 
         property_type, listing_type, created_date, last_scraped_date,  
         bedrooms, bathrooms, max_guests, listing_url)

# write_csv(property_new_select, "./data/airdna/clean/property_new_select.csv")
saveRDS(property_new_select, "./data/airdna/clean/property_new_select.Rds")

# rm(property_new)
```


<!-- ------------------------------------------------------------ --> 
<!-- ------------------------------------------------------------ --> 

# Monthly

```{r, include=FALSE}
monthly_new_raw <- read_csv("./data/airdna/raw/Australia_Airbnb_Monthly_2019-03-22.zip")

# any(duplicated(monthly_new_raw$`property_new ID`))
# length(unique(monthly_new_raw$`property_new ID`))
# any(duplicated(monthly_new_raw$`Host ID`))
# length(unique(monthly_new_raw$`Host ID`))
```

Raw dataset consists of `r nrow(monthly_new_raw)` records of `monthly_new` statistics of Austrlian Airbnb properties scraped by Airdna. 

## General

```{r, include=FALSE}
# temp <- filter(monthly_new_raw, `Property ID` == 1476936)

monthly_new <- monthly_new_raw %>% 
  clean_names() %>%
  remove_empty(c("rows", "cols")) %>% 
  select(-country, -state, -city, -neighborhood) %>% 
  mutate(latitude = ifelse(property_id == 1476936, latitude*-1, latitude)) %>% 
  arrange(property_id, reporting_month)

# # temp <- 
# monthly_new %>%
#   select(property_id, latitude, longitude) %>%
#   group_by(property_id) %>%
#   summarise(latitude = first(latitude),
#             longitude = first(longitude)) %>%
#   arrange(property_id) %>%
#   write_csv("./data/airdna/clean/property_new_geo.csv")

monthly_new %<>% left_join(SA1_2016_join_new)

# table(is.na(SA1_2016_join_new$SA1_MAIN16))
# table(is.na(monthly_new$SA1_MAIN16))

rm(monthly_new_raw, SA1_2016_join_new)
gc()
```

* Dots\spaces in the names were replaced by (single) underscores.
* One location with wrong coordinates (`property_id` == 1476936) was fixed.
* Completely empty variables (`Zipcode` and `Metropolitan_Statistical_Area`) were dropped.
* Variable `Country` was droped. 
* Variables `State`, `City` and `Neighborhood` were droped (no clue how assigned; will be added from ABS data linked spatially by coordinates)

Dataset consists of `r nrow(monthly_new)` records in `r ncol(monthly_new)` variables:

```{r}
names(monthly_new)
```

## Data structure

```{r, results='asis'}
# skim(monthly_new)
skim(monthly_new) %>% skimr::kable()
```

## Missing data

```{r}
gg_miss_var(monthly_new, show_pct = TRUE)
```

## Reporting Month

```{r, fig.width=8, fig.height=8, out.width="800px", out.height="800px"}
# tab(property_new, property_type)
ggplot(monthly_new, aes(reporting_month)) + 
  geom_bar() + coord_flip() + theme_minimal()
```

Clearly not all propoerties have all months!

Number of months property has data on: 

```{r}
monthly_new %>% 
  group_by(property_id) %>% 
  summarise(Months = n()) %>% 
  ungroup() %>% 
  ggplot(aes(Months)) + 
  geom_bar() + theme_minimal()
```

**Peaks at certain periods?**

## Active

```{r, fig.width=4, fig.height=3, out.width="400px", out.height="300px"}
ggplot(monthly_new, aes(active))+ 
  geom_bar() + theme_minimal()
```

## Occupancy Rate

```{r, fig.width=4, fig.height=3, out.width="400px", out.height="300px"}
monthly_new %>% 
  select(occupancy_rate) %>% 
  summarytools::descr(stats = c("mean", "med", "min", "max")) 

ggplot(monthly_new, aes(occupancy_rate)) +
  geom_histogram() + theme_minimal()
```

## `revenue_native`

```{r, fig.width=4, fig.height=3, out.width="400px", out.height="300px"}
monthly_new %>% 
  select(revenue_native) %>% 
  summarytools::descr(stats = c("mean", "med", "min", "max")) 

# monthly_new %>% 
#   # filter(revenue_native > 0) %>% 
#   # filter(revenue_native < 1000000) %>% 
#   ggplot(aes(`Revenue_(Native)`)) +
#   geom_histogram() + theme_minimal()
```


```{r}
# write_csv(monthly_new, "./Data/Airbnb/clean/monthly_new.csv")
saveRDS(monthly_new, "./data/airdna/clean/monthly_new.Rds")

monthly_new_select <- monthly_new %>% 
  select(property_id, reporting_month, active, 
         listing_type, property_type, bedrooms, 
         number_of_reservations, occupancy_rate, reservation_days, 
         available_days, blocked_days,
         revenue_usd, revenue_native, adr_usd, adr_native,
         starts_with("SA"), starts_with("STE"))

# write_csv(monthly_new_select, "./data/airdna/clean/monthly_new_select.csv")
saveRDS(monthly_new_select, "./data/airdna/clean/monthly_new_select.Rds")

# rm(monthly_new)
```
