==> analyses/19_Airbnb_seasonal_SA3.Rmd <==
---
title: "Airbnb - seasonality of GNO"
subtitle: "Monthly relative on SA3 level"
# author: "Radoslaw Panczak"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: 
  html_document: 
    highlight: pygments
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 4
---

<!-- ------------------------------------------------------------ --> 
<!-- ------------------------------------------------------------ --> 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE, fig.width=6, fig.height=3.75, dpi=300, out.width="600px", out.height="375px")

knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
set.seed(12345)
options(scipen = 999)

library(pacman) 
p_load(tidyverse, magrittr, lubridate, scales, 
       sf, tmap, tmaptools, 
       sjmisc, kableExtra, 
       kml)

tmap_mode("view") # makes map interactive
```

```{r geodata_load, include=FALSE}
# SA3
SA3 <- readRDS(file = "./data/geo/SA3_2016_AUST_clean.rds") %>% 
  mutate(SA3_CODE16 = as.numeric(SA3_CODE16), 
         STE_CODE16 = as.numeric(STE_CODE16)) %>% 
  select(-starts_with("SA4"), -starts_with("GCC"), -AREASQKM16)

SA3_centr <- readRDS(file = "./data/geo/SA3_centr.rds") %>% 
  mutate(SA3_CODE16 = as.numeric(SA3_CODE16))

# STATES
STE <- readRDS(file = "./data/geo/STE_2016_AUST_clean.rds") %>%  
  mutate(STE_CODE16 = as.numeric(as.character(STE_CODE16))) %>% 
  filter(STE_CODE16 != 9)

# qtm(STE)
```

<!-- ------------------------------------------------------------ --> 
<!-- ------------------------------------------------------------ --> 

# Data 

```{r include=FALSE}
airbnb_sa3 <- readRDS(file = "./data/airdna/clean/airbnb_sa3.rds")
airbnb_sa3_18 <- readRDS(file = "./data/airdna/clean/airbnb_sa3_18.rds")

# length(unique(SA3$SA3_CODE16))
# length(unique(airbnb_monthly$SA3_CODE16))
```

Absolute numbers

```{r}
ggplot(data=airbnb_sa3, 
       aes(x=reporting_month, y=max_gno/1000, 
           color=STE_SHORT, group=SA3_CODE16)) +
  geom_line(alpha=0.65) +
  theme_minimal() + xlab("") + ylab("Airbnb max_gno [*1,000]")
```

Relative GNO were calculated monthly for 2018:

```{r}
ggplot(airbnb_sa3_18, 
       aes(x=reporting_month, y=share_gno, 
           color=STE_SHORT, group=SA3_CODE16)) +
  geom_line(alpha=0.5) +
  theme_minimal() + xlab("") + ylab("Airbnb relative max_gno {%}")

```


```{r eval=FALSE, include=FALSE}
# This is the view across regions:

temp <- left_join(SA3, airbnb_sa3_18) %>% 
  filter(!is.na(share_gno)) %>% 
  mutate(reporting_month = month(reporting_month))#, label = TRUE, abbr = TRUE))

tmap_mode("plot") 

tm_shape(temp) +
  tm_polygons(col = "share_gno", n = 5, style = "quantile", palette = "YlGnBu") +
  tm_facets(by = "reporting_month") +
  tm_shape(STE) +
  tm_borders("white", lwd=2)

# tmap_save(tmap_last(), "plot/airbnb_sa3_18_seasonal.png", width=1920, height=1080, asp=0)

tmap_mode("view") 

rm(temp)
```


<!-- ------------------------------------------------------------ --> 
<!-- ------------------------------------------------------------ --> 

# Clustering temporal trajectories

## Methods

See Genolini *et al.* (2015) at https://www.jstatsoft.org/article/view/v065i04 

```{r, include=FALSE}
# kml needs wide data
airbnb_sa3_18_wide <- airbnb_sa3_18 %>% 
  select(SA3_CODE16, SA3_NAME16, reporting_month, share_gno) %>% 
  mutate(Month = lubridate::month(reporting_month, label = TRUE, abbr = TRUE)) %>% 
  select(-reporting_month) %>% 
  spread(Month, share_gno) %>% 
  select(SA3_CODE16, SA3_NAME16, Jan:Dec)

cld_airbnb <- cld(airbnb_sa3_18_wide, timeInData = 3:14, idAll = airbnb_sa3_18_wide$SA3_CODE16)

# kml(cld_airbnb, nbRedraw = 2, toPlot = "both")

kml(cld_airbnb, nbClusters=2:10)

airbnb_sa3_18_wide$cluster3 <- getClusters(cld_airbnb, 3)
airbnb_sa3_18_wide$cluster9 <- getClusters(cld_airbnb, 9)

airbnb_sa3_18 %<>% left_join(select(airbnb_sa3_18_wide, SA3_CODE16, starts_with("cluster")))

SA3 %<>% filter(SA3_NAME16 != "Illawarra Catchment Reserve") %>% 
  left_join(select(airbnb_sa3_18_wide, SA3_CODE16, starts_with("cluster")))

file.remove("cld_airbnb.Rdata")

regions <- readRDS(file = "./data/airdna/clean/regions.rds") %>% 
  select(sa3, region_types) %>% 
  rename(SA3_NAME16 = sa3) %>% 
  left_join(airbnb_sa3_18)
```

## Number of clusters

No clear evidence for optimal amount of clusters

```{r}
# choice(cld_airbnb)
plotAllCriterion(cld_airbnb)
```

## Different cluster solutions

### 3 clusters

#### Relative and absolute numbers

```{r}
airbnb_sa3_18 %>% 
  mutate(month = strftime(reporting_month, "%m")) %>% 
  ggplot(aes(x=month, y=share_gno)) +
  # geom_line(aes(color=cluster3, group=SA3_CODE16), alpha=0.20) +
  geom_smooth(aes(group=cluster3, color=cluster3), size = 1, se = FALSE) +
  theme_minimal() + xlab("") + ylab("Airbnb max_gno {%}") +
  scale_colour_brewer(palette = "Dark2") 

airbnb_sa3_18 %>% 
  mutate(month = strftime(reporting_month, "%m")) %>% 
  ggplot(aes(x=month, y=max_gno/1000)) +
  geom_col(aes(fill=cluster3)) +
  theme_minimal() + xlab("") + ylab("Airbnb max_gno [*1,000]")+ 
  scale_fill_brewer(palette = "Dark2") 

sjmisc::frq(SA3$cluster3)

tm_shape(SA3) +
  tm_polygons("cluster3", border.col = "white", lwd = 0.5, 
              id = "SA3_NAME16", popup.vars = c("cluster3"), 
              palette = "Dark2") +
  tm_shape(STE) +
  tm_borders(lwd = 0.5, alpha = 0.5)  
```

#### Selected regions

```{r}
regions %>% 
  mutate(month = strftime(reporting_month, "%m")) %>% 
  ggplot(aes(x=month, y=share_gno)) +
  geom_line(aes(color=cluster3, group=SA3_CODE16), size=1) +
  theme_minimal() + xlab("") + ylab("Airbnb max_gno {%}") +
  scale_colour_brewer(palette = "Dark2") +
  facet_wrap(vars(region_types))

SA3 %>% 
  right_join(select(regions, SA3_NAME16)) %>% 
  tm_shape() +
  tm_polygons("cluster3", border.col = "white", lwd = 0.5, 
              id = "SA3_NAME16", popup.vars = c("cluster3"), 
              palette = "Dark2") +
  tm_shape(STE) +
  tm_borders(lwd = 0.5, alpha = 0.5)  
```

### 9 clusters

#### Relative and absolute numbers

```{r}
airbnb_sa3_18 %>% 
  mutate(month = strftime(reporting_month, "%m")) %>% 
  ggplot(aes(x=month, y=share_gno)) +
  # geom_line(aes(color=cluster9, group=SA3_CODE16), alpha=0.20) +
  geom_smooth(aes(group=cluster9, color=cluster9), size = 1, se = FALSE) +
  theme_minimal() + xlab("") + ylab("Airbnb max_gno {%}") +
  scale_colour_brewer(palette = "Set3") 

airbnb_sa3_18 %>% 
  mutate(month = strftime(reporting_month, "%m")) %>% 
  ggplot(aes(x=month, y=max_gno/1000)) +
  geom_col(aes(fill=cluster9)) +
  theme_minimal() + xlab("") + ylab("Airbnb max_gno [*1,000]")+ 
  scale_fill_brewer(palette = "Set3") 

sjmisc::frq(SA3$cluster9)

tm_shape(SA3) +
  tm_polygons("cluster9", border.col = "white", lwd = 0.5, 
              id = "SA3_NAME16", popup.vars = c("cluster9"), 
              palette = "Set3") +
  tm_shape(STE) +
  tm_borders(lwd = 0.5, alpha = 0.5)  

tmap_mode("plot") 

tm_shape(SA3) +
  tm_polygons(col = "black", border.col = "white", lwd = 0.5) +
  tm_facets(by = "cluster9", free.coords = FALSE, ncol = 3) +
  tm_shape(STE) +
  tm_borders(lwd = 0.5, alpha = 0.5)

tmap_mode("view") 
```

#### Selected regions

```{r}
regions %>% 
  mutate(month = strftime(reporting_month, "%m")) %>% 
  ggplot(aes(x=month, y=share_gno)) +
  geom_line(aes(color=cluster9, group=SA3_CODE16), size=1) +
  theme_minimal() + xlab("") + ylab("Airbnb max_gno {%}") +
  scale_colour_brewer(palette = "Set3") +
  facet_wrap(vars(region_types))

SA3 %>% 
  right_join(select(regions, SA3_NAME16)) %>% 
  tm_shape() +
  tm_polygons("cluster9", border.col = "white", lwd = 0.5, 
              id = "SA3_NAME16", popup.vars = c("cluster9"), 
              palette = "Set3") +
  tm_shape(STE) +
  tm_borders(lwd = 0.5, alpha = 0.5)  
```


# Geofacets of max_gno

Using 9 clusters solution

## Absolute

```{r}
p_load(geofacet)

airbnb_sa3 %<>% 
  left_join(select(airbnb_sa3_18_wide, SA3_CODE16, starts_with("cluster"))) 

# frq(airbnb_sa3, STE_SHORT)
# frq(airbnb_sa3, STE_SHORT)
```

```{r}
airbnb_ste <- airbnb_sa3 %>% 
  group_by(STE_SHORT, reporting_month, cluster9) %>% 
  summarise(max_gno = sum(max_gno)) %>% 
  ungroup() %>% 
  group_by(STE_SHORT, cluster9) %>% arrange(reporting_month) %>% mutate(cumulative = cumsum(max_gno)) %>% 
  ungroup() %>% 
  arrange(STE_SHORT, cluster9, reporting_month)

airbnb_ste %>% 
  ggplot(aes(reporting_month, max_gno/1000)) +
  scale_fill_brewer(palette = "Set3") + 
  scale_x_date(date_labels = "%y/%m") + 
  geom_col(aes(fill = cluster9)) +
  facet_geo(~ STE_SHORT, grid = "aus_grid1", scales = "free_y") +
  geom_text(aes(label = STE_SHORT), size = 3, col = "grey40", x = -Inf, y = Inf, hjust = -0.5, vjust = 1.2) +
  labs(
    title = "Monthly Airbnb max_gno in Australia",
    # caption = "Data Source: Airdna",
    y = "max_gno [*1,000]", x = "") +
  theme_light() +
  theme(strip.background = element_blank(),
        strip.text = element_blank(),
        legend.position = "none")

```

## Relative

```{r}
airbnb_ste <- airbnb_sa3 %>% 
  filter(reporting_month >= ymd("2018-01-01")) %>% 
  filter(reporting_month <= ymd("2018-12-01")) %>% 
  group_by(STE_SHORT, reporting_month, cluster9) %>% 
  summarise(max_gno = sum(max_gno)) %>% 
  ungroup() %>% 
  group_by(STE_SHORT) %>% 
  mutate(reservations_total = sum(as.numeric(max_gno))) %>% 
  ungroup() %>% 
  mutate(share_gno = (max_gno/reservations_total)*100) %>% 
  mutate(share_gno = ifelse(max_gno==0 & reservations_total==0, 0, share_gno)) %>% 
  arrange(STE_SHORT, cluster9, reporting_month) 

airbnb_ste %>% 
  ggplot(aes(reporting_month, share_gno)) +
  scale_fill_brewer(palette = "Set3") + 
  geom_col(aes(fill = cluster9)) +
  geom_text(aes(label = STE_SHORT), size = 3, col = "grey40", x = -Inf, y = Inf, hjust = -0.5, vjust = 1.2) +
  facet_geo(~ STE_SHORT, grid = "aus_grid1") +
  labs(
    title = "Monthly relative Airbnb max_gno in Australia",
    # caption = "Data Source: Airdna",
    y = "% of yearly revenue", x = "") +
  theme_light() +
  theme(strip.background = element_blank(),
        strip.text = element_blank(),
        legend.position = "none")
```


```{r}
# Session info
# sessionInfo()
```
