==> analyses/26_Airbnb_SA2_SEIFA.Rmd <==
---
title: "Airbnb"
subtitle: "Rents & income on SA2 level"
author: "Radoslaw Panczak"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: 
  html_document: 
    highlight: pygments
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 4
    toc_float: yes
---

<!-- ------------------------------------------------------------ --> 
<!-- ------------------------------------------------------------ --> 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE, 
                      fig.width=8, fig.height=6, 
                      out.width="800px", out.height="600px", 
                      dpi=300)

knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
set.seed(12345)
options(scipen = 999)

library(pacman) 
p_load(tidyverse, magrittr, lubridate, scales, readxl, janitor, imputeTS,
       sf, tmap, tmaptools, 
       ggridges, skimr, summarytools, sjmisc, kableExtra, naniar)

tmap_mode("view") # makes map interactive
```

<!-- ------------------------------------------------------------ --> 

# SA2 areas

```{r geodata_load, include=FALSE}
# SA2
SA2 <- readRDS(file = "./data/geo/SA2_2016_AUST_clean.rds") %>% 
  # filter(SA4_CODE16 != "901") %>% 
  filter(! SA2_5DIG16 %in% c("91001", "91002", "91004", "11161")) %>% 
  select(-starts_with("SA3")) %>% 
  select(-starts_with("SA1")) %>% 
  select(-starts_with("GCC")) %>% 
  select(-AREASQKM16) %>% 
  mutate(STE_CODE16 = as.numeric(as.character(STE_CODE16))) %>% 
  mutate(SA2_MAIN16 = as.numeric(as.character(SA2_MAIN16))) %>% 
  mutate(SA2_5DIG16 = as.numeric(as.character(SA2_5DIG16))) %>% 
  mutate(SA4_CODE16 = as.numeric(as.character(SA4_CODE16))) %>% 
  mutate(SA4_GCC_CODE16 = case_when(
    SA4_CODE16 %in% c(402, 403, 404) ~ 401,
    SA4_CODE16 %in% c(302:305, 310, 311, 313, 314) ~ 301,
    SA4_CODE16 %in% c(207:214) ~ 206,
    SA4_CODE16 %in% c(503:507) ~ 502,
    SA4_CODE16 %in% c(115:128 ) ~ 102,
    TRUE ~ SA4_CODE16
  )
  ) %>% 
  mutate(SA4_GCC_NAME16 = case_when(
    SA4_GCC_CODE16 == 401 ~ "Greater Adelaide",
    SA4_GCC_CODE16 == 301 ~ "Greater Brisbane",
    SA4_GCC_CODE16 == 206 ~ "Greater Melbourne",
    SA4_GCC_CODE16 == 502 ~ "Greater Perth",
    SA4_GCC_CODE16 == 102 ~ "Greater Sydney",
    TRUE ~ SA4_NAME16
  )
  ) %>% 
  select(-SA4_CODE16, -SA4_NAME16)

# centroids from arcgis - with option 'inside'
SA2_centr <- st_read("./data/geo/1270055001_sa2_2016_aust_shape/SA2_2016_AUST_centr_inside.shp",
                     stringsAsFactors = FALSE) %>% 
  # filter(SA4_CODE16 != "901") %>% 
  filter(! SA2_5DIG16 %in% c("91001", "91002", "91004", "11161")) %>% 
  select(-starts_with("SA3")) %>% 
  select(-starts_with("SA1")) %>% 
  select(-starts_with("GCC")) %>% 
  select(-AREASQKM16) %>% 
  mutate(STE_CODE16 = as.numeric(as.character(STE_CODE16))) %>% 
  mutate(SA2_MAIN16 = as.numeric(as.character(SA2_MAIN16))) %>% 
  mutate(SA2_5DIG16 = as.numeric(as.character(SA2_5DIG16))) %>% 
  mutate(SA4_CODE16 = as.numeric(as.character(SA4_CODE16))) %>% 
  mutate(SA4_GCC_CODE16 = case_when(
    SA4_CODE16 %in% c(402, 403, 404) ~ 401,
    SA4_CODE16 %in% c(302:305, 310, 311, 313, 314) ~ 301,
    SA4_CODE16 %in% c(207:214) ~ 206,
    SA4_CODE16 %in% c(503:507) ~ 502,
    SA4_CODE16 %in% c(115:128 ) ~ 102,
    TRUE ~ SA4_CODE16
  )
  ) %>% 
  mutate(SA4_GCC_NAME16 = case_when(
    SA4_GCC_CODE16 == 401 ~ "Greater Adelaide",
    SA4_GCC_CODE16 == 301 ~ "Greater Brisbane",
    SA4_GCC_CODE16 == 206 ~ "Greater Melbourne",
    SA4_GCC_CODE16 == 502 ~ "Greater Perth",
    SA4_GCC_CODE16 == 102 ~ "Greater Sydney",
    TRUE ~ SA4_NAME16
  )
  ) %>% 
  select(-SA4_CODE16, -SA4_NAME16)

# STATES
STE <- readRDS(file = "./data/geo/STE_2016_AUST_clean.rds") %>%  
  mutate(STE_CODE16 = as.numeric(as.character(STE_CODE16))) %>% 
  filter(STE_CODE16 != 9) 

# qtm(STE)
```

Starting point is set of `r number(nrow(SA2))` areas. This excludes 4 remote areas - three from Other Territories (but keeps Jervis Bay!) and one from NSW (Lord Howe Island)

```{r}
# Christmas Island, Cocos (Keeling) Islands, Norfolk Island, Lord Howe Island

readRDS(file = "./data/geo/SA2_2016_AUST_clean.rds") %>% 
  filter(SA2_5DIG16 %in% c("91001", "91002", "91004", "11161")) %>% 
  select(SA2_NAME16, STE_NAME16) %>% 
  st_drop_geometry()

# readRDS(file = "./data/geo/SA2_2016_AUST_clean.rds") %>% 
#   filter(STE_NAME16 == "Other Territories") %>% 
#   select(SA2_NAME16, STE_NAME16) %>% 
#   st_drop_geometry()
# 
# readRDS(file = "./data/geo/SA2_2016_AUST_clean.rds") %>% 
#   filter(SA2_NAME16 == "Norfolk Island") %>% 
#   # select(SA2_NAME16, STE_NAME16) %>% 
#   st_drop_geometry()
```


<!-- ------------------------------------------------------------ --> 

# Airbnb data

Data prep focusing on monthly income generated on SA2 level. 

```{r, include=FALSE}
monthly_new_select <- readRDS("./data/airdna/clean/monthly_new_select.Rds") 

summary(monthly_new_select$reporting_month)
```

Raw datasets consist of `r comma(length(unique(monthly_new_select$property_id)))` properties and 
`r comma(nrow(monthly_new_select))` monthly stats. 

## Selection

`r comma(nrow(filter(monthly_new_select, reporting_month < ymd("2016-01-01"))))` data points from before 2016-01-01 were excluded.

```{r, include=FALSE}
monthly_new_prep_revenue <- monthly_new_select %>% 
  filter(reporting_month >= ymd("2016-01-01"))

# length(unique(monthly_new_prep_revenue$property_id)) 
# length(unique(temp1$property_id))

rm(monthly_new_select)
gc()
```

```{r, include=FALSE}
#### Other Territories
OT <- monthly_new_prep_revenue %>% 
  filter(SA2_5DIG16 %in% c("91001", "91002", "91004", "11161")) 

monthly_new_prep_revenue %<>% 
  filter(! SA2_5DIG16 %in% c("91001", "91002", "91004", "11161")) 

#### zero aud - convert to usd
zero_aud <- monthly_new_prep_revenue %>% 
  dplyr::filter(revenue_native == 0 & revenue_usd > 0)

# sjmisc::descr(zero_aud, revenue_usd)

#### missing aud - convert to usd
mis_aud <- monthly_new_prep_revenue %>% 
  dplyr::filter(is.na(revenue_native) & !is.na(revenue_usd))

# sjmisc::descr(mis_aud, revenue_usd)

monthly_new_prep_revenue %<>% 
  mutate(revenue_native = ifelse(test = revenue_native == 0 & revenue_usd > 0, 
                                 yes = revenue_usd*1.4, 
                                 no = revenue_native)) %>% 
  mutate(revenue_native = ifelse(test = is.na(revenue_native) & !is.na(revenue_usd), 
                                 yes = revenue_usd*1.4, 
                                 no = revenue_native)) 
```

`r comma(nrow(OT))` records of `r comma(length(unique(OT$property_id)))` properties located in Other Teritories were excluded. 

Next, `r format(nrow(zero_aud), nsmall=0, big.mark=",")` monthly records of `r comma(length(unique(zero_aud$property_id)))` properties that had zero AUD income but some information in USD had income recalculated using 1.4 exchange rate.

Next, `r format(nrow(mis_aud), nsmall=0, big.mark=",")` monthly records of `r comma(length(unique(mis_aud$property_id)))` properties that had missing AUD income but some information in USD had income recalculated using 1.4 exchange rate.

```{r, include=FALSE}
rm(zero_aud, mis_aud, OT)
```

## Outliers

There are few ouliers in terms of monthly income

```{r}
temp <- monthly_new_prep_revenue %>% 
  arrange(desc(revenue_native)) %>% 
  # select(property_id, reporting_month, reservation_days, 
  #        starts_with("Revenue"), starts_with("ADR")) %>% 
  slice(1:36)

monthly_new_prep_revenue %>% 
  arrange(desc(revenue_native)) %>% 
  select(property_id, reporting_month, reservation_days,
         starts_with("Revenue"), starts_with("ADR")) %>% 
  slice(1:36)

# View(filter(property_select, property_id == 14999741))
# Luxury 80ft Motor Yacht - 4 Bedroom
# $6,990.00 per night

# View(filter(property_select, property_id == 30109331))
# Kalinya Estate - Homestead & Lodge -26 p
# $3,200.00 per night
```

From the top 11 monthly income values (with income above $90k) data corrected using USD values.

One very high value was replaced to average ADR * booked days from previous & next month from the same property.


```{r include=FALSE}
monthly_new_prep_revenue %<>% 
  mutate(revenue_native = ifelse(test = revenue_native > 95000, 
                                 yes = revenue_usd*1.4, no = revenue_native))

monthly_new_prep_revenue %<>% 
  mutate(revenue_native = ifelse(test = property_id == 10045109 & reporting_month == "2016-02-01", 
                                 yes = mean(c(45.42, 48.11))*9, no = revenue_native))

monthly_new_prep_revenue %<>% 
  mutate(revenue_native = ifelse(test = property_id == 11287673, 
                                 yes = reservation_days*28, no = revenue_native))
```

Nine very high monthly values were recalculated from more realistic ADR (in USD) indicators.

```{r include=FALSE}
monthly_new_prep_revenue %<>% 
  mutate(revenue_native = ifelse(test = property_id == 23709347	&
                                   reporting_month == ymd("2018-04-01"), 
                                 yes = reservation_days*adr_usd*1.4, no = revenue_native))

monthly_new_prep_revenue %<>% 
  mutate(revenue_native = ifelse(test = property_id == 22267809	&
                                   reporting_month == ymd("2018-05-01"), 
                                 yes = reservation_days*adr_usd*1.4, no = revenue_native))

monthly_new_prep_revenue %<>% 
  mutate(revenue_native = ifelse(test = property_id == 22536123	&
                                   reporting_month == ymd("2018-07-01"), 
                                 yes = reservation_days*adr_usd*1.4, no = revenue_native))
monthly_new_prep_revenue %<>% 
  mutate(revenue_native = ifelse(test = property_id == 22536123	&
                                   reporting_month == ymd("2018-04-01"), 
                                 yes = reservation_days*adr_usd*1.4, no = revenue_native))

monthly_new_prep_revenue %<>% 
  mutate(revenue_native = ifelse(test = property_id == 11926837	&
                                   reporting_month == ymd("2016-06-01"), 
                                 yes = reservation_days*adr_usd*1.4, no = revenue_native))

monthly_new_prep_revenue %<>% 
  mutate(revenue_native = ifelse(test = property_id == 27554784	&
                                   reporting_month == ymd("2019-02-01"), 
                                 yes = reservation_days*adr_usd*1.4, no = revenue_native))


monthly_new_prep_revenue %<>% 
  mutate(revenue_native = ifelse(test = property_id == 15206751	&
                                   reporting_month == ymd("2017-04-01"), 
                                 yes = reservation_days*adr_usd*1.4, no = revenue_native))

monthly_new_prep_revenue %<>% 
  mutate(revenue_native = ifelse(test = property_id == 15206751	&
                                   reporting_month == ymd("2017-05-01"), 
                                 yes = reservation_days*adr_usd*1.4, no = revenue_native))

monthly_new_prep_revenue %<>% 
  mutate(revenue_native = ifelse(test = property_id == 27554784	&
                                   reporting_month == ymd("2019-01-01"), 
                                 yes = reservation_days*adr_usd*1.4, no = revenue_native))
```

```{r include=FALSE}
saveRDS(monthly_new_prep_revenue, "./data/airdna/clean/monthly_new_prep_revenue.rds")

monthly_new_prep_revenue <- readRDS("./data/airdna/clean/monthly_new_prep_revenue.rds")
```

Final dataset consists of `r comma(nrow(monthly_new_prep_revenue))`
monthly records of `r comma(length(unique(monthly_new_prep_revenue$property_id)))` properties in `r comma(length(unique(monthly_new_prep_revenue$SA2_MAIN16)))` SA2 areas. 

## SA2 areas

Locations of Airbnb properties (points) were linked to ABS data of SA2 areas from 2016. Spatial join in ArcGIS was used with `CLOSEST` option to capture locations that did not overlap with polygons (see example here  https://www.airbnb.com/rooms/19103554 - due to privacy reasons, locations are not exact). 

`r nrow(filter(SA2, !SA2_MAIN16 %in% unique(monthly_new_prep_revenue$SA2_MAIN16)))` SA2s (out of `r nrow(SA2)`) have no Airbnbs inside. 

```{r}
temp <- SA2 %>% 
  filter(!SA2_MAIN16 %in% unique(monthly_new_prep_revenue$SA2_MAIN16))

temp %>% 
  select(-SA2_MAIN16, -SA2_5DIG16, -STE_CODE16, -SA4_GCC_CODE16) %>% 
  tm_shape() +
  tm_polygons(col = "red", alpha = 0.5)
```

**These areas were kept in analyses, unless excluded for reasons further down.**

```{r include=FALSE, eval=FALSE}
# SA2 %<>% 
#   filter(!SA2_MAIN16 %in% temp$SA2_MAIN16)
# 
# SA2_centr %<>% 
#   filter(!SA2_MAIN16 %in% temp$SA2_MAIN16)
# 
# rm(temp)

```

## SA2 prepared dataset

Data for Brisbane City SA2. Individual properties with smoothed trend:

```{r}
monthly_new_prep_revenue %>% 
  filter(SA2_5DIG16 %in% c(31105)) %>% 
  ggplot(aes(x=reporting_month, y=revenue_native)) +
  geom_line(aes(group=property_id), alpha = .33) +
  geom_smooth(size = 1, se = FALSE) +
  theme_minimal() + xlab("") + ylab("Revenue AUD")
```

This dataset gets aggregated for each month taking revenue from all properties. It then looks like that:

```{r}
monthly_new_prep_revenue %>% 
  filter(SA2_5DIG16 %in% c(31105)) %>% 
  group_by(reporting_month) %>% 
  summarise(revenue_native = sum(revenue_native, na.rm = TRUE)) %>% 
  ggplot(aes(x=reporting_month, y=revenue_native)) +
  geom_col() +
  theme_minimal() + xlab("") + ylab("Revenue AUD")
```

#### 'Filling' time series

Areas where no income was reported during specific reporting month (including areas with no Airbnbs at all!) had those months filled with 0 AUD.  

```{r, include=FALSE}
temp <- monthly_new_prep_revenue %>% 
  group_by(SA2_MAIN16, reporting_month) %>% 
  summarise(revenue = sum(revenue_native, na.rm = TRUE))

airbnb_sa2 <- left_join(st_drop_geometry(SA2), temp) %>% 
  expand(SA2_MAIN16, reporting_month) %>% 
  filter(!is.na(reporting_month)) %>% 
  left_join(st_drop_geometry(SA2)) %>% 
  left_join(temp) %>% 
  tidyr::replace_na(list(revenue = 0)) %>% 
  mutate(revenue = as.integer(round(revenue))) %>% 
  mutate_if(is.factor, as.character) %>% 
  ungroup() %>% 
  group_by(SA2_MAIN16) %>% arrange(reporting_month) %>% mutate(cumulative = cumsum(revenue)) %>% 
  ungroup() %>% 
  arrange(SA2_MAIN16, reporting_month)

# summarytools::freq(airbnb_sa2$reporting_month)
stopifnot(nrow(SA2) * length(unique(monthly_new_prep_revenue$reporting_month)) == nrow(airbnb_sa2))

# table(is.na(monthly_new_prep_revenue$SA1_7DIG16))

rm(temp)
gc()
```

Result is in the form of 'income curves' with `r nrow(airbnb_sa2)` data points of Airbnb revenue for `r comma(length(unique(airbnb_sa2$SA2_MAIN16)))` areas. 

#### Absolute monthly 

```{r}
ggplot(data=airbnb_sa2, 
       aes(x=reporting_month, y=revenue/1000000, 
           color=STE_NAME16, group=SA2_MAIN16)) +
  geom_line(alpha=0.5) +
  theme_minimal() + xlab("") + ylab("Airbnb revenue [*1,000,000]$")
```

#### Cumulative values

```{r}
ggplot(data=airbnb_sa2, 
       aes(x=reporting_month, y=cumulative/1000000, 
           color=STE_NAME16, group=SA2_MAIN16)) +
  geom_line(alpha=0.5) +
  theme_minimal() + xlab("") + ylab("Airbnb revenue [*1,000,000]$")
```


<!-- ------------------------------------------------------------ --> 

# SEIFA indices

```{r}
SEIFA <- read_xls("data/ABS/SEIFA/raw/2033055001 - sa2 indexes.xls", 
                  sheet = "Table 1", skip = 5, na = "-") %>% 
  clean_names() %>%
  remove_empty(c("rows", "cols")) %>% 
  dplyr::rename(SA2_MAIN16 = `x1`, SA2_NAME16 = `x2`, 
                IRSD_s = `score_3`, IRSD_d = `decile_4`,
                IRSAD_s = `score_5`, IRSAD_d = `decile_6`, 
                IER_s = `score_7`, IER_d = `decile_8`, 
                IEO_s = `score_9`, IEO_d = `decile_10`, 
                URP = `x11`) %>% 
  dplyr::mutate(IRSD_d = as.integer(IRSD_d),
                IRSAD_d = as.integer(IRSAD_d),
                IER_d = as.integer(IER_d),
                IEO_d = as.integer(IEO_d)) %>% 
  filter(! SA2_NAME16 %in% c("Christmas Island", 
                             "Cocos (Keeling) Islands", 
                             "Norfolk Island", 
                             "Lord Howe Island")) # Other Teritories out, but Jervis Bay is in!!!


# missing all seifa
SEIFA_mis_sf <- SA2 %>% 
  anti_join(select(SEIFA, SA2_MAIN16, URP))

# checking for airbnbs
airbnb_SEIFA_mis <- monthly_new_prep_revenue %>%
  filter(SA2_MAIN16 %in% unique(SEIFA_mis_sf$SA2_MAIN16))


# missing one of seifa
SEIFA_na <- anti_join(SEIFA, select(drop_na(SEIFA), SA2_MAIN16))

SEIFA_na_sf <- SA2 %>% 
  right_join(SEIFA_na)

# checking for airbnbs
airbnb_SEIFA_na <- monthly_new_prep_revenue %>%
  filter(SA2_MAIN16 %in% unique(SEIFA_na$SA2_MAIN16))
```

### Areas with no SEIFA

`r nrow(SEIFA_mis_sf)` areas, usually non-residential or with no/low pop. 

There are `r comma(nrow(airbnb_SEIFA_mis))` records of `r comma(length(unique(airbnb_SEIFA_mis$property_id)))` propoerties in these areas. 

```{r}
SEIFA_mis_sf %>% 
  select(-SA2_MAIN16, -SA2_5DIG16, -STE_CODE16, -SA4_GCC_CODE16) %>% 
  tm_shape() +
  tm_polygons(col = "red", alpha = 0.5, lwd = 1)

SA2 %<>%
  filter(!SA2_MAIN16 %in% SEIFA_mis_sf$SA2_MAIN16)

SA2_centr %<>%
  filter(!SA2_MAIN16 %in% SEIFA_mis_sf$SA2_MAIN16)

airbnb_sa2 %<>%
  filter(!SA2_MAIN16 %in% SEIFA_mis_sf$SA2_MAIN16)
```

### Areas without one or more indices

`r nrow(SEIFA_na)` areas, usually with small pops. **Excluded from analyses!**.

There are `r number(nrow(airbnb_SEIFA_na))` records of `r number(length(unique(airbnb_SEIFA_na$property_id)))` propoerties in these areas. 

```{r}
SEIFA_na_sf %>% 
  select(-SA2_MAIN16, -SA2_5DIG16, -STE_CODE16, -SA4_GCC_CODE16) %>% 
  tm_shape() +
  tm_polygons(col = "red", alpha = 0.5, lwd = 1)

SA2 %<>%
  filter(!SA2_MAIN16 %in% SEIFA_na_sf$SA2_MAIN16)

SA2_centr %<>%
  filter(!SA2_MAIN16 %in% SEIFA_na_sf$SA2_MAIN16)

airbnb_sa2 %<>%
  filter(!SA2_MAIN16 %in% SEIFA_na_sf$SA2_MAIN16)
```

```{r include=FALSE}
rm(SEIFA_na, SEIFA_na_sf, SEIFA_mis_sf, airbnb_SEIFA_na, airbnb_SEIFA_mis)
gc()
```

```{r}
airbnb_sa2 %<>% left_join(select(SEIFA, SA2_MAIN16, ends_with("_d")))
```

### IRSD

Index of Relative Socio-economic Disadvantage deciles 

```{r}
SEIFA_sa2_sf <- left_join(SA2, SEIFA)
rm(SEIFA)

tm_shape(SEIFA_sa2_sf) +
  tm_polygons(col = "IRSD_d", n = 10, palette = "BrBG", alpha = 0.5, lwd = 0)
```

### IRSAD

Index of Relative Socio-economic Advantage and Disadvantage deciles

```{r}
tm_shape(SEIFA_sa2_sf) +
  tm_polygons(col = "IRSAD_d", n = 10, palette = "BrBG", alpha = 0.5, lwd = 0)
```

### IER

Index of Economic Resources deciles

```{r}
tm_shape(SEIFA_sa2_sf) +
  tm_polygons(col = "IER_d", n = 10, palette = "BrBG", alpha = 0.5, lwd = 0)
```

### IEO

Index of Education and Occupation deciles

```{r}
tm_shape(SEIFA_sa2_sf) +
  tm_polygons(col = "IEO_d", n = 10, palette = "BrBG", alpha = 0.5, lwd = 0)
```

<!-- ------------------------------------------------------------ --> 

# Explanatory variables

## Remotness

Data are from *1270.0.55.005 - Australian Statistical Geography Standard (ASGS): Volume 5 - Remoteness Structure, July 2016* [info](https://www.abs.gov.au/AUSSTATS/abs@.nsf/DetailsPage/1270.0.55.005July%202016?OpenDocument).

`No usual address` and `Migratory - Offshore - Shipping` categories are excluded. 

Data are on SA1 level. Population of SA1 was used to find majority classification within each SA2.

```{r include=FALSE}
remote <- read_csv("data/ABS/1270055005_ra_2016_aust_csv/RA_2016_AUST.zip", 
                   col_types = cols(AREA_ALBERS_SQKM = col_skip(), 
                                    MB_CODE_2016 = col_skip(), RA_CODE_2016 = col_integer(), 
                                    STATE_CODE_2016 = col_skip(), STATE_NAME_2016 = col_skip())) %>% 
  rename(SA1_MAIN16 = SA1_MAINCODE_2016,
         SA1_7DIG16 = SA1_7DIGITCODE_2016) %>% 
  group_by(SA1_MAIN16) %>% 
  filter(row_number()==1) %>% 
  ungroup() %>% 
  filter(!str_detect(RA_NAME_2016, "Migratory - Offshore - Shipping")) %>% 
  filter(!str_detect(RA_NAME_2016, "No usual address"))

pop_sa1 <- read_csv("data/ABS/2016 Census GCP Statistical Area 1 for AUST/2016Census_G01_AUS_SA1.csv.gz") %>% 
  select(SA1_7DIGITCODE_2016, Tot_P_P) %>% 
  rename(SA1_7DIG16 = SA1_7DIGITCODE_2016,
         pop_16 = Tot_P_P)

SA1 <- st_read("./data/geo/1270055001_sa1_2016_aust_shape/SA1_2016_AUST.shp", 
               stringsAsFactors = FALSE) %>% 
  st_drop_geometry() %>% 
  select(SA1_MAIN16, SA2_MAIN16, SA2_5DIG16, SA2_NAME16) %>% 
  mutate(SA1_MAIN16 = as.numeric(as.character(SA1_MAIN16))) %>% 
  mutate(SA2_MAIN16 = as.numeric(as.character(SA2_MAIN16))) %>% 
  mutate(SA2_5DIG16 = as.numeric(as.character(SA2_5DIG16)))

remote %<>% 
  left_join(pop_sa1) %>% 
  left_join(SA1)

remote %<>% 
  rename(pop_sa1_16 = pop_16) %>% 
  group_by(SA2_MAIN16) %>% 
  mutate(pop_sa2_16 = sum(pop_sa1_16)) %>% 
  ungroup() 
```

Original data (SA1 level!):

```{r}
remote %>% 
  frq(RA_NAME_2016)
```

SA2 assigned to majority

```{r}
remote_ra <- 
  remote %>% 
  group_by(SA2_MAIN16, RA_CODE_2016) %>% 
  mutate(pop_sa2_16_cat = sum(pop_sa1_16)) %>% 
  ungroup() %>% 
  mutate(pop_sa2_16_s = pop_sa2_16_cat/pop_sa2_16) %>% 
  arrange(SA2_MAIN16, desc(pop_sa2_16_s)) %>% 
  group_by(SA2_MAIN16, RA_CODE_2016) %>%
  filter(row_number() == 1) %>% 
  ungroup() %>% 
  select(-SA1_MAIN16, -SA1_7DIG16, -pop_sa1_16) %>% 
  select(SA2_MAIN16, SA2_5DIG16, SA2_NAME16, everything())

remote_ra_one <- 
  remote_ra %>% 
  group_by(SA2_MAIN16) %>%
  filter(row_number() == 1) %>% 
  ungroup() %>% 
  filter(! SA2_5DIG16 %in% c("91001", "91002", "91004", "11161"))
```

Share of majority category

```{r}
remote_ra_one %>% 
  descr(pop_sa2_16_s)

ggplot(remote_ra_one, aes(pop_sa2_16_s)) + 
  geom_histogram() +
  ylab("Number of SA2s") + xlab("Share of majority category") + theme_minimal()
```

Map 

```{r}
SA2 %>% 
  left_join(remote_ra_one) %>% 
  tm_shape() + 
  tm_polygons(col = "RA_NAME_2016", alpha = .5, border.col = NA, id = "SA3_NAME16")
```


## Level of urbanization

Data are from *1270.0.55.001 - Australian Statistical Geography Standard (ASGS): Volume 1 - Main Structure and Greater Capital City Statistical Areas, July 2016* [info](https://www.abs.gov.au/AUSSTATS/abs@.nsf/Lookup/1270.0.55.001Main+Features10018July%202016?OpenDocument)

`No usual address` and `Migratory - Offshore - Shipping` categories are excluded. 

Data are on SA1 level. Population of SA1 was used to find majority classification within each SA3.


```{r include=FALSE}
urban <- read_csv("data/ABS/1270055004_sa1_ucl_sosr_sos_2016_aust_csv/raw/SA1_UCL_SOSR_SOS_2016_AUST.zip", 
                  col_types = cols(AREA_ALBERS_SQKM = col_skip(), 
                                   SOSR_CODE_2016 = col_integer(), 
                                   SOS_CODE_2016 = col_integer(), 
                                   STE_CODE_2016 = col_skip(), STE_NAME_2016 = col_skip())) %>% 
  rename(SA1_MAIN16 = SA1_MAINCODE_2016,
         SA1_7DIG16 = SA1_7DIGITCODE_2016) %>% 
  select(-starts_with("UCL")) %>% 
  filter(!str_detect(SOSR_NAME_2016, "Migratory - Offshore - Shipping")) %>% 
  filter(!str_detect(SOSR_NAME_2016, "No usual address"))

urban %<>% 
  left_join(pop_sa1) %>% 
  left_join(SA1)

urban %<>% 
  rename(pop_sa1_16 = pop_16) %>% 
  group_by(SA2_MAIN16) %>% 
  mutate(pop_sa2_16 = sum(pop_sa1_16)) %>% 
  ungroup() 

rm(pop_sa1, SA1)
gc()
```

### Section of State Range (SOSR) 

Original data

```{r}
urban %>% 
  frq(SOSR_NAME_2016)
```

SA2 assigned to majority

```{r}
urban_sosr <- 
  urban %>% 
  select(-SOS_CODE_2016, -SOS_NAME_2016) %>% 
  group_by(SA2_MAIN16, SOSR_CODE_2016) %>% 
  mutate(pop_sa2_16_cat = sum(pop_sa1_16)) %>% 
  ungroup() %>% 
  mutate(pop_sa2_16_s = pop_sa2_16_cat/pop_sa2_16) %>% 
  arrange(SA2_MAIN16, desc(pop_sa2_16_s)) %>% 
  group_by(SA2_MAIN16, SOSR_CODE_2016) %>%
  filter(row_number() == 1) %>% 
  ungroup() %>% 
  select(-SA1_MAIN16, -SA1_7DIG16, -pop_sa1_16) %>% 
  select(SA2_MAIN16, SA2_5DIG16, SA2_NAME16, everything())

urban_sosr_one <- 
  urban_sosr %>% 
  group_by(SA2_MAIN16) %>%
  filter(row_number() == 1) %>% 
  ungroup() %>% 
  filter(! SA2_5DIG16 %in% c(90101, 90102, 90104, 10803))
```

Share of majority category

```{r}
urban_sosr_one %>% 
  descr(pop_sa2_16_s)

ggplot(urban_sosr_one, aes(pop_sa2_16_s)) + 
  geom_histogram() +
  ylab("Number of SA3") + xlab("Share of majority category") + theme_minimal()
```

Map 

```{r}
SA2 %>% 
  left_join(urban_sosr_one) %>% 
  tm_shape() + 
  tm_polygons(col = "SOSR_NAME_2016", alpha = .5, border.col = NA, id = "SA3_NAME16")
```

### Section of State (SOS)

Original data

```{r}
urban %>% 
  frq(SOS_NAME_2016)
```

SA2 assigned to majority

```{r}
urban_sos <- 
  urban %>% 
  select(-SOSR_CODE_2016, -SOSR_NAME_2016) %>% 
  group_by(SA2_MAIN16, SOS_CODE_2016) %>% 
  mutate(pop_sa2_16_cat = sum(pop_sa1_16)) %>% 
  ungroup() %>% 
  mutate(pop_sa2_16_s = pop_sa2_16_cat/pop_sa2_16) %>% 
  arrange(SA2_MAIN16, desc(pop_sa2_16_s)) %>% 
  group_by(SA2_MAIN16, SOS_CODE_2016) %>%
  filter(row_number() == 1) %>% 
  ungroup() %>% 
  select(-SA1_MAIN16, -SA1_7DIG16, -pop_sa1_16) %>% 
  select(SA2_MAIN16, SA2_5DIG16, SA2_NAME16, everything())

urban_sos_one <- 
  urban_sos %>% 
  group_by(SA2_MAIN16) %>%
  filter(row_number() == 1) %>% 
  ungroup() %>% 
  filter(! SA2_5DIG16 %in% c(90101, 90102, 90104, 10803))
```

Share of majority category

```{r}
urban_sos_one %>% 
  descr(pop_sa2_16_s)

ggplot(urban_sos_one, aes(pop_sa2_16_s)) + 
  geom_histogram() +
  ylab("Number of SA2s") + xlab("Share of majority category") + theme_minimal()
```

Map 

```{r}
SA2 %>% 
  left_join(urban_sos_one) %>% 
  tm_shape() + 
  tm_polygons(col = "SOS_NAME_2016", alpha = .5, border.col = NA, id = "SA3_NAME16")
```

## ABS Census 2016 data 

### HOSD Housing Suitability

*This variable is new to 2016 and is a measure of housing utilisation based on a comparison of the number of bedrooms in a dwelling with a series of household demographics, such as the number of usual residents, their relationship to each other, age and sex. The criteria are based on the Canadian National Occupancy Standard (CNOS). This variable can be used to identify if a dwelling is either under or over utilised.* 

(Info)[https://www.abs.gov.au/ausstats/abs@.nsf/Lookup/by%20Subject/2900.0~2016~Main%20Features~HOSD%20Household%20Suitability~10123]

```{r}
HOSD <- read_xlsx("data/ABS/misc/raw/SA2 by HOSD Housing Suitability.xlsx", skip = 8) %>% 
  clean_names() %>%
  remove_empty(c("rows", "cols")) %>% 
  dplyr::rename(SA2_NAME16 = `x2`) %>% 
  filter(!grepl("Migratory - Offshore - Shipping", SA2_NAME16)) %>% 
  filter(!grepl("Cells in this table", SA2_NAME16)) %>% 
  filter(! SA2_NAME16 %in% c("Christmas Island", "Cocos (Keeling) Islands", 
                             "Jervis Bay", "Norfolk Island", 
                             "SA2", "Total")) %>%  
  dplyr::mutate(four_needed = four_or_more_extra_bedrooms_needed / total,
                three_needed = three_extra_bedrooms_needed / total,
                two_needed = two_extra_bedrooms_needed / total,
                one_needed = one_extra_bedroom_needed / total,
                optimal = no_bedrooms_needed_or_spare / total,
                one_spare = one_bedroom_spare / total,
                two_spare = two_bedrooms_spare / total,
                three_spare = three_bedrooms_spare / total,
                four_spare = four_or_more_bedrooms_spare / total) %>% 
  dplyr::mutate(needed = (four_or_more_extra_bedrooms_needed  + 
                            three_extra_bedrooms_needed + 
                            two_extra_bedrooms_needed + 
                            one_extra_bedroom_needed) / total,
                spare = (one_bedroom_spare +
                           two_bedrooms_spare + 
                           three_bedrooms_spare + 
                           four_or_more_bedrooms_spare) / total)

# missing 
HOSD_sf <- SA2 %>% 
  left_join(HOSD)

# checking for airbnbs
airbnb_HOSD_mis <- monthly_new_prep_revenue %>%
  filter(! SA2_NAME16 %in% HOSD$SA2_NAME16) %>% 
  distinct(property_id, SA2_NAME16) %>% 
  group_by(SA2_NAME16) %>% 
  summarise(count = n())
rm(airbnb_HOSD_mis)

```

```{r}
HOSD_sf %>% 
  tm_shape() +
  tm_polygons(c("needed", "spare"), 
              n = 7, palette = "RdYlBu", id = "SA3_NAME16") +
  tm_facets(free.scales.fill = FALSE)
```

### Other census indicators???

## Distance to coast

### Binary

Binary indicator has been created for costal proximity. SA2s that touch the coast have been flagged. Note: this is very crude proxy. Doesn't work well for large areas and in the areas with river estuaries.  

```{r}
coast_bin_sa2 <- read_csv("./data/airdna/raw/Spatial_join/coast_bin_sa2.zip",
                          col_types = cols(COAST = col_integer())) %>% 
  select(SA2_MAIN16, COAST) %>% 
  rename(coast_bin = COAST)

airbnb_sa2 %<>% left_join(coast_bin_sa2)

SA2 <- left_join(SA2, coast_bin_sa2)

rm(coast_bin_sa2)
```

```{r}
filter(airbnb_sa2, reporting_month == as.Date("2017-10-01")) %>% 
  frq(coast_bin)
```

```{r}
tm_shape(SA2) +
  tm_polygons(col = "coast_bin", n = 2, alpha = 0.5, lwd = 0.5)
```

### Average distance

Alternative solution is to calculate distance to coast for all properties and then average that over area. However that leaves the areas without properties without data.

```{r}
coast_near <- read_csv("./data/airdna/raw/Spatial_join/coast_near.zip") %>% 
  rename(property_id = Property_ID, coast_dist = NEAR_DIST) %>% 
  select(property_id, coast_dist)

coast_near %>% 
  select(coast_dist) %>% 
  summarytools::descr(stats = c("mean", "med", "min", "max")) 

SA1_2016_join_new <- readRDS("./data/airdna/clean/SA1_2016_join_new.Rds") %>% 
  dplyr::select(property_id, SA2_5DIG16)

coast_near <- left_join(coast_near, SA1_2016_join_new) %>% 
  group_by(SA2_5DIG16) %>% 
  summarise(cost_mean = mean(coast_dist),
            cost_median = median(coast_dist))

coast_near %>% 
  select(cost_mean, cost_median) %>% 
  summarytools::descr(stats = c("mean", "med", "min", "max")) 

SA2 <- left_join(SA2, coast_near)

rm(coast_near, SA1_2016_join_new)
```

```{r}
tm_shape(SA2) +
  tm_polygons(col = "cost_median", n = 5, alpha = 0.5, lwd = 0.5)
```

## Tourist accomodation

ABS dataset '8635.0 - Tourist Accommodation, Australia, 2015-16' [dataset](http://www.abs.gov.au/AUSSTATS/abs@.nsf/Lookup/8635.0Main+Features12015-16?OpenDocument) could be used to provide information on supply and deman of accomodation on SA2 level. Several indicators are available on **monthly basis for July 2015 - June 2016 period**:

Establishments 
Rooms 
Bed spaces 
Room nights occupied 
Room nights available 
Room occupancy rate
Guest nights occupied 
Guest nights available 
Bed occupancy rate
Takings from accommodation 
Average takings per room night occupied
Average takings per room night available

See external description for an example of 'Guest Nights Occupied' calculations.

```{r}
saveRDS(airbnb_sa2, file = "./data/airdna/clean/airbnb_sa2.rds")

# write_csv(airbnb_sa2, path = "./data/airdna/clean/airbnb_sa2.csv", 
#           na = "NA", append = FALSE, col_names = TRUE)

```

<!-- ------------------------------------------------------------ --> 
<!-- ------------------------------------------------------------ --> 

# Session info

```{r}
sessionInfo()
```
