==> analyses/31_Airbnb_SA2_APM_rents-model_unit_INLA.Rmd <==
---
title: "Airbnb - relation between rent & Airbnb revenue and supply on SA2 level"
subtitle: "Models"
author: "Radoslaw Panczak"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: 
  html_document: 
    highlight: pygments
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 4
    toc_float: yes
---

<!-- ------------------------------------------------------------ --> 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, 
                      message = TRUE, 
                      warning = TRUE, 
                      cache = FALSE,
                      fig.width=8, fig.height=6, 
                      out.width="800px", out.height="600px", 
                      dpi=300)

knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
set.seed(12345)
options(scipen = 999)

library(pacman) 
p_load(tidyverse, magrittr, lubridate, scales,
       summarytools, sjmisc, kableExtra
)

```


```{r include=FALSE}
APM_sa2 <- readRDS("data/APM/clean/APM_sa2.Rds") %>% 
  mutate(adr_25p_int = 7*adr_25p_int,
         adr_50p_int = 7*adr_50p_int,
         adr_75p_int = 7*adr_75p_int) %>% 
  mutate(adr_25p_int = as.integer(round(adr_25p_int)),
         adr_50p_int = as.integer(round(adr_50p_int)),
         adr_75p_int = as.integer(round(adr_75p_int)))

```

<!-- ------------------------------------------------------------ --> 

# Selection

```{r include=FALSE}
hist(APM_sa2$record_count)
summary(APM_sa2$record_count)

hist(APM_sa2$n_listings)
summary(APM_sa2$n_listings)

APM_sa2_cor <- 
  APM_sa2 %>%
  filter(!is.na(rent_50p_int)) %>%
  filter(!is.na(adr_50p_int)) %>%
  group_by(SA2_MAIN16, SA2_NAME16, property_type) %>% 
  summarize(
    min_apm = min(record_count, na.rm = TRUE),
    min_airbnb = min(n_listings, na.rm = TRUE),
    mean_apm = mean(record_count, na.rm = TRUE),
    mean_airbnb = mean(n_listings, na.rm = TRUE),
    cor_25 = cor(rent_25p_int, adr_25p_int),
    cor_50 = cor(rent_50p_int, adr_50p_int),
    cor_75 = cor(rent_75p_int, adr_75p_int)
  ) %>% 
  ungroup() %>% 
  filter(min_apm >= 25) %>%
  filter(min_airbnb >= 10)

summary(APM_sa2_cor$mean_airbnb)
summary(APM_sa2_cor$mean_apm)

summary(APM_sa2_cor$min_airbnb)
summary(APM_sa2_cor$min_apm)

```

<!-- ------------------------------------------------------------ --> 

# Data

**Houses, WA & NSW only**

```{r}
data <- 
  select(APM_sa2_cor, SA2_MAIN16, property_type) %>% 
  filter(property_type == "Unit") %>% 
  left_join(APM_sa2) %>% 
  # filter(STE_NAME16 == "New South Wales") %>%
  # filter(STE_NAME16 == "Western Australia") %>%
  # filter(STE_NAME16 %in% c("New South Wales", "Western Australia")) %>%
  group_by(SA2_MAIN16) %>% 
  mutate(time = row_number()) %>% 
  ungroup() %>% 
  select(-property_type)

data %>% select(SA2_MAIN16) %>% distinct() %>% nrow()

length(unique(data$SA2_MAIN16))

```

<!-- ------------------------------------------------------------ --> 

# Testing INLA

```{r echo=TRUE, results='hide'}
# https://becarioprecario.bitbucket.io/inla-gitbook/ch-multilevel.html#multilevel-models-for-longitudinal-data

p_load(INLA, INLAutils)

# ####
# Jeffreys prior 
a1 <- 5e-5
b1 <- 5e-5
lgprior1 <- list(prec = list(param = c(a1, b1)))

# Gelman prior
a2 <- -0.5
b2 <- 5e-5
lgprior2 <- list(prec = list(param = c(a2, b2)))

# iid prior 
# Schrödle & Held 2010 & Blangiardo et al 2013
a0 <- 1
b0 <- 0.1
prior.nu <- list(prec = list(param = c(a0, b0)))

# intercept & fixed
inla.set.control.fixed.default() 

# intercept ~ N(0,0) 
# other fixed effects ~ N(0, 0.001) 
# 
# where the format is N(mean, precision) 
# precision = inverse of the variance. 

# PC prior
U <- 1
hyper.prec = list(theta = list(
  prior = "pc.prec",
  param = c(U, 0.01)
))

# scaling
inla.setOption(scale.model.default = TRUE)

```

## random region, no adr

```{r}
f0 <- rent_50p_int ~ 1 + time +
  f(STE_NAME16, model = "iid", constr=TRUE) + 
  f(SA2_NAME16, model = "iid", constr=TRUE) 

m0 <- inla(f0, data = data,
           control.compute = list(dic= TRUE, waic = TRUE))

summary(m0)
# plot(m0)
# autoplot(m0)

# m0_sa2 <- m0$summary.random$SA2_NAME16
m0_fit <- m0$summary.fitted.values

m0_fit$SA2_NAME16 <- data$SA2_NAME16
m0_fit$STE_NAME16 <- data$STE_NAME16
m0_fit$x <- data$time
m0_fit$rent_50p_int <- data$rent_50p_int
m0_fit$n_listings <- data$n_listings

ggplot(m0_fit, aes(x = rent_50p_int)) +
  geom_line(aes(y = rent_50p_int), col = "darkorchid4") +
  geom_point(aes(y = `0.5quant`, size = n_listings), alpha = 0.3) +
  facet_wrap(~STE_NAME16) +
  theme_minimal() + ylab("Predicted") + xlab("Rent")

m0_fit %>% 
  filter(SA2_NAME16 %in% c("Augusta", "Trigg - North Beach - Watermans Bay")) %>% 
  ggplot(aes(x = x)) +
  geom_ribbon(aes(ymin = `0.025quant`, ymax = `0.975quant`, group = SA2_NAME16), alpha=0.3) +
  geom_line(aes(y = `0.5quant`, group = SA2_NAME16), color = "purple") +
  geom_point(aes(y = rent_50p_int, col = SA2_NAME16, size = n_listings), 
             alpha = 0.5) +
  theme_minimal() + ylab("Rent") + xlab("Time")
```

## random region

```{r}
f1 <- rent_50p_int ~ 1 + adr_50p_int + time +
  f(STE_NAME16, model = "iid", constr=TRUE) + 
  f(SA2_NAME16, model = "iid", constr=TRUE) 

m1 <- inla(f1, data = data,
           control.compute = list(dic= TRUE, waic = TRUE))

summary(m1)
# plot(m1)
# autoplot(m1)

# m1_sa2 <- m1$summary.random$SA2_NAME16
m1_fit <- m1$summary.fitted.values

m1_fit$SA2_NAME16 <- data$SA2_NAME16
m1_fit$STE_NAME16 <- data$STE_NAME16
m1_fit$x <- data$time
m1_fit$rent_50p_int <- data$rent_50p_int
m1_fit$n_listings <- data$n_listings

ggplot(m1_fit, aes(x = rent_50p_int)) +
  geom_line(aes(y = rent_50p_int), col = "darkorchid4") +
  geom_point(aes(y = `0.5quant`, size = n_listings), alpha = 0.3) +
  facet_wrap(~STE_NAME16) +
  theme_minimal() + ylab("Predicted") + xlab("Rent")

m1_fit %>% 
  filter(SA2_NAME16 %in% c("Augusta", "Trigg - North Beach - Watermans Bay")) %>% 
  ggplot(aes(x = x)) +
  geom_ribbon(aes(ymin = `0.025quant`, ymax = `0.975quant`, group = SA2_NAME16), alpha=0.3) +
  geom_line(aes(y = `0.5quant`, group = SA2_NAME16, col = SA2_NAME16)) +
  geom_point(aes(y = rent_50p_int, col = SA2_NAME16, size = n_listings), 
             alpha = 0.5) +
  theme_minimal() + ylab("Rent") + xlab("Time")
```


## random time region

```{r}
f2 <- rent_50p_int ~ 1 + adr_50p_int +
  f(STE_NAME16, model = "iid", constr=TRUE) + 
  f(SA2_NAME16, time, model = "iid", constr=TRUE) 

m2 <- inla(f2, data = data,
           control.compute = list(dic= TRUE, waic = TRUE))

summary(m2)
# plot(m2)
# autoplot(m2)

# m2_sa2 <- m2$summary.random$SA2_NAME16
m2_fit <- m2$summary.fitted.values

m2_fit$SA2_NAME16 <- data$SA2_NAME16
m2_fit$STE_NAME16 <- data$STE_NAME16
m2_fit$x <- data$time
m2_fit$rent_50p_int <- data$rent_50p_int
m2_fit$n_listings <- data$n_listings

ggplot(m2_fit, aes(x = rent_50p_int)) +
  geom_line(aes(y = rent_50p_int), col = "darkorchid4") +
  geom_point(aes(y = `0.5quant`, size = n_listings), alpha = 0.3) +
  facet_wrap(~STE_NAME16) +
  theme_minimal() + ylab("Predicted") + xlab("Rent")

m2_fit %>% 
  filter(SA2_NAME16 %in% c("Augusta", "Trigg - North Beach - Watermans Bay")) %>% 
  ggplot(aes(x = x)) +
  geom_ribbon(aes(ymin = `0.025quant`, ymax = `0.975quant`, group = SA2_NAME16), alpha=0.3) +
  geom_line(aes(y = `0.5quant`, group = SA2_NAME16, col = SA2_NAME16)) +
  geom_point(aes(y = rent_50p_int, col = SA2_NAME16, size = n_listings), 
             alpha = 0.5) +
  theme_minimal() + ylab("Rent") + xlab("Time")
```


## random region, smooth adr

```{r}
f3 <- rent_50p_int ~ 1 + time +
  f(adr_50p_int, model = "rw1", hyper = hyper.prec) +
  f(STE_NAME16, model = "iid", constr=TRUE) + 
  f(SA2_NAME16, model = "iid", constr=TRUE) 

m3 <- inla(f3, data = data,
           control.compute = list(dic= TRUE, waic = TRUE))

summary(m3)
# plot(m3)
# autoplot(m3)

# m3_sa2 <- m3$summary.random$SA2_NAME16
m3_fit <- m3$summary.fitted.values

m3_fit$SA2_NAME16 <- data$SA2_NAME16
m3_fit$STE_NAME16 <- data$STE_NAME16
m3_fit$x <- data$time
m3_fit$rent_50p_int <- data$rent_50p_int
m3_fit$n_listings <- data$n_listings

m3_fit$adr_50p_int <- data$adr_50p_int
m3_fit$record_count <- data$record_count

ggplot(m3_fit, aes(x = rent_50p_int)) +
  geom_line(aes(y = rent_50p_int), col = "darkorchid4") +
  geom_point(aes(y = `0.5quant`, size = n_listings), alpha = 0.3) +
  facet_wrap(~STE_NAME16) +
  theme_minimal() + ylab("Predicted") + xlab("Rent")

m3_fit %>% 
  filter(SA2_NAME16 %in% c("Augusta", "Trigg - North Beach - Watermans Bay")) %>% 
  ggplot(aes(x = x)) +
  geom_ribbon(aes(ymin = `0.025quant`, ymax = `0.975quant`, group = SA2_NAME16), alpha=0.3) +
  geom_line(aes(y = `0.5quant`, group = SA2_NAME16, col = SA2_NAME16)) +
  geom_point(aes(y = rent_50p_int, col = SA2_NAME16, size = n_listings), 
             alpha = 0.5) +
  theme_minimal() + ylab("Rent") + xlab("Time")
```


## random region time, smooth adr

```{r}
f4 <- rent_50p_int ~ 1 + 
  f(adr_50p_int, model = "rw1", hyper = hyper.prec) +
  f(STE_NAME16, model = "iid", constr=TRUE) + 
  f(SA2_NAME16, time, model = "iid", constr=TRUE) 

m4 <- inla(f4, data = data,
           control.compute = list(dic= TRUE, waic = TRUE))

summary(m4)
# plot(m4)
# autoplot(m4)

# m4_sa2 <- m4$summary.random$SA2_NAME16
m4_fit <- m4$summary.fitted.values

m4_fit$SA2_NAME16 <- data$SA2_NAME16
m4_fit$STE_NAME16 <- data$STE_NAME16
m4_fit$x <- data$time
m4_fit$rent_50p_int <- data$rent_50p_int
m4_fit$n_listings <- data$n_listings

ggplot(m4_fit, aes(x = rent_50p_int)) +
  geom_line(aes(y = rent_50p_int), col = "darkorchid4") +
  geom_point(aes(y = `0.5quant`, size = n_listings), alpha = 0.3) +
  facet_wrap(~STE_NAME16) +
  theme_minimal() + ylab("Predicted") + xlab("Rent")

m4_fit %>% 
  filter(SA2_NAME16 %in% c("Augusta", "Trigg - North Beach - Watermans Bay")) %>% 
  ggplot(aes(x = x)) +
  geom_ribbon(aes(ymin = `0.025quant`, ymax = `0.975quant`, group = SA2_NAME16), alpha=0.3) +
  geom_line(aes(y = `0.5quant`, group = SA2_NAME16, col = SA2_NAME16)) +
  geom_point(aes(y = rent_50p_int, col = SA2_NAME16, size = n_listings), 
             alpha = 0.5) +
  theme_minimal() + ylab("Rent") + xlab("Time")
```


## random region, smooth adr, smooth time

```{r}
f5 <- rent_50p_int ~ 1 + 
  f(time, model = "rw1", hyper = hyper.prec) +
  f(adr_50p_int, model = "rw1", hyper = hyper.prec) +
  f(STE_NAME16, model = "iid", constr=TRUE) + 
  f(SA2_NAME16, model = "iid", constr=TRUE) 

m5 <- inla(f5, data = data,
           control.compute = list(dic= TRUE, waic = TRUE))

summary(m5)
# plot(m5)
# autoplot(m5)

# m5_sa2 <- m5$summary.random$SA2_NAME16
m5_fit <- m5$summary.fitted.values

m5_fit$SA2_NAME16 <- data$SA2_NAME16
m5_fit$STE_NAME16 <- data$STE_NAME16
m5_fit$x <- data$time
m5_fit$rent_50p_int <- data$rent_50p_int
m5_fit$n_listings <- data$n_listings

m5_fit$adr_50p_int <- data$adr_50p_int
m5_fit$record_count <- data$record_count

ggplot(m5_fit, aes(x = rent_50p_int)) +
  geom_line(aes(y = rent_50p_int), col = "darkorchid4") +
  geom_point(aes(y = `0.5quant`, size = n_listings), alpha = 0.3) +
  facet_wrap(~STE_NAME16) +
  theme_minimal() + ylab("Predicted") + xlab("Rent")

m5_fit %>% 
  filter(SA2_NAME16 %in% c("Augusta", "Trigg - North Beach - Watermans Bay")) %>% 
  ggplot(aes(x = x)) +
  geom_ribbon(aes(ymin = `0.025quant`, ymax = `0.975quant`, group = SA2_NAME16), alpha=0.3) +
  geom_line(aes(y = `0.5quant`, group = SA2_NAME16, col = SA2_NAME16)) +
  geom_point(aes(y = rent_50p_int, col = SA2_NAME16, size = n_listings), 
             alpha = 0.5) +
  # geom_point(aes(y = adr_50p_int, col = SA2_NAME16, size = record_count), 
  #            alpha = 0.5) +
  theme_minimal() + ylab("Rent") + xlab("Time")

```

## Comparison

```{r}
dotchart(c(m0$dic$dic, m1$dic$dic, m2$dic$dic, 
           m3$dic$dic, m4$dic$dic, m5$dic$dic),
         labels = c("mod0", "mod1", "mod2", "mod3", "mod4", "mod5"), 
         main = "DIC")

dotchart(c(m0$waic$waic, m1$waic$waic, m2$waic$waic, 
           m3$waic$waic, m4$waic$waic, m5$waic$waic),
         labels = c("mod0", "mod1", "mod2", "mod3", "mod4", "mod5"), 
         main = "WAIC")

```

```{r}
m3_fit$diff <- (m3_fit$rent_50p_int - m3_fit$`0.5quant`) / m3_fit$rent_50p_int
m3_ste <- m3$summary.random$STE_NAME16
m3_adr <- m3$summary.random$adr_50p_int

m3_adr %>% 
  ggplot(aes(x = ID)) +
  geom_ribbon(aes(ymin = `0.025quant`, ymax = `0.975quant`), alpha=0.3) +
  geom_line(aes(y = `0.5quant`)) +
  theme_minimal() + ylab("Rent") + xlab("ADR")

m3_fit %>% 
  filter(SA2_NAME16 %in% c("Hobart")) %>% 
  ggplot(aes(x = x)) +
  geom_ribbon(aes(ymin = `0.025quant`, ymax = `0.975quant`, group = SA2_NAME16), alpha=0.3) +
  geom_line(aes(y = `0.5quant`, group = SA2_NAME16, col = SA2_NAME16)) +
  geom_point(aes(y = rent_50p_int, col = SA2_NAME16, size = n_listings),
             alpha = 0.5) +
  # geom_point(aes(y = adr_50p_int, col = SA2_NAME16, size = record_count),
  #            alpha = 0.5) +
  theme_minimal() + ylab("Rent") + xlab("Time")

```

```{r}
m5_fit$diff <- (m5_fit$rent_50p_int - m5_fit$`0.5quant`) / m5_fit$rent_50p_int
m5_ste <- m5$summary.random$STE_NAME16
m5_adr <- m5$summary.random$adr_50p_int

m5_adr %>% 
  ggplot(aes(x = ID)) +
  geom_ribbon(aes(ymin = `0.025quant`, ymax = `0.975quant`), alpha=0.3) +
  geom_line(aes(y = `0.5quant`)) +
  theme_minimal() + ylab("Rent") + xlab("ADR")

m5_fit %>% 
  filter(SA2_NAME16 %in% c("Hobart")) %>% 
  ggplot(aes(x = x)) +
  geom_ribbon(aes(ymin = `0.025quant`, ymax = `0.975quant`, group = SA2_NAME16), alpha=0.3) +
  geom_line(aes(y = `0.5quant`, group = SA2_NAME16, col = SA2_NAME16)) +
  geom_point(aes(y = rent_50p_int, col = SA2_NAME16, size = n_listings),
             alpha = 0.5) +
  # geom_point(aes(y = adr_50p_int, col = SA2_NAME16, size = record_count), 
  #            alpha = 0.5) +
  theme_minimal() + ylab("Rent") + xlab("Time")

```



sessionInfo()
```
